tweaks to generation + clipping
