+++ Spectrogram/GeneratedColorMaps.hpp
@@ -1,2136 +1,13 @@
-//
-//Machine generated color maps extracted from matplotlib cm module
-//
+// Copyright (c) 2016-2016 Josh Blum
+// SPDX-License-Identifier: BSL-1.0
 
 #include <string>
 #include <vector>
+#include <utility>
+#include <cstddef>
 
-static std::vector<std::vector<double>> make_color_map_Blues(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.96862745285,0.984313726425,1.0});
-    out.push_back(std::vector<double>{0.125,0.870588243008,0.921568632126,0.96862745285});
-    out.push_back(std::vector<double>{0.25,0.776470601559,0.858823537827,0.937254905701});
-    out.push_back(std::vector<double>{0.375,0.61960786581,0.792156875134,0.882352948189});
-    out.push_back(std::vector<double>{0.5,0.419607847929,0.68235296011,0.839215695858});
-    out.push_back(std::vector<double>{0.625,0.258823543787,0.572549045086,0.776470601559});
-    out.push_back(std::vector<double>{0.75,0.129411771894,0.443137258291,0.709803938866});
-    out.push_back(std::vector<double>{0.875,0.0313725508749,0.317647069693,0.611764729023});
-    out.push_back(std::vector<double>{1.0,0.0313725508749,0.188235297799,0.419607847929});
-    return out;
-}
+//! Get a listing of all title, name pairs for each color map
+std::vector<std::pair<const std::string, const std::string>> available_color_maps(void);
 
-static std::vector<std::vector<double>> make_color_map_BuGn(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.96862745285,0.988235294819,0.992156863213});
-    out.push_back(std::vector<double>{0.125,0.898039221764,0.960784316063,0.976470589638});
-    out.push_back(std::vector<double>{0.25,0.800000011921,0.92549020052,0.901960790157});
-    out.push_back(std::vector<double>{0.375,0.600000023842,0.847058832645,0.78823530674});
-    out.push_back(std::vector<double>{0.5,0.40000000596,0.760784327984,0.643137276173});
-    out.push_back(std::vector<double>{0.625,0.254901975393,0.68235296011,0.46274510026});
-    out.push_back(std::vector<double>{0.75,0.137254908681,0.54509806633,0.270588248968});
-    out.push_back(std::vector<double>{0.875,0.0,0.427450984716,0.172549024224});
-    out.push_back(std::vector<double>{1.0,0.0,0.266666680574,0.105882354081});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_BuPu(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.96862745285,0.988235294819,0.992156863213});
-    out.push_back(std::vector<double>{0.125,0.878431379795,0.92549020052,0.956862747669});
-    out.push_back(std::vector<double>{0.25,0.749019622803,0.827450990677,0.901960790157});
-    out.push_back(std::vector<double>{0.375,0.61960786581,0.737254917622,0.854901969433});
-    out.push_back(std::vector<double>{0.5,0.549019634724,0.588235318661,0.776470601559});
-    out.push_back(std::vector<double>{0.625,0.549019634724,0.419607847929,0.694117665291});
-    out.push_back(std::vector<double>{0.75,0.533333361149,0.254901975393,0.615686297417});
-    out.push_back(std::vector<double>{0.875,0.505882382393,0.0588235296309,0.486274510622});
-    out.push_back(std::vector<double>{1.0,0.301960796118,0.0,0.29411765933});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_GnBu(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.96862745285,0.988235294819,0.941176474094});
-    out.push_back(std::vector<double>{0.125,0.878431379795,0.952941179276,0.858823537827});
-    out.push_back(std::vector<double>{0.25,0.800000011921,0.921568632126,0.772549033165});
-    out.push_back(std::vector<double>{0.375,0.658823549747,0.866666674614,0.709803938866});
-    out.push_back(std::vector<double>{0.5,0.482352942228,0.800000011921,0.768627464771});
-    out.push_back(std::vector<double>{0.625,0.305882364511,0.701960802078,0.827450990677});
-    out.push_back(std::vector<double>{0.75,0.168627455831,0.549019634724,0.745098054409});
-    out.push_back(std::vector<double>{0.875,0.0313725508749,0.407843142748,0.674509823322});
-    out.push_back(std::vector<double>{1.0,0.0313725508749,0.250980407,0.505882382393});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Greens(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.96862745285,0.988235294819,0.960784316063});
-    out.push_back(std::vector<double>{0.125,0.898039221764,0.960784316063,0.878431379795});
-    out.push_back(std::vector<double>{0.25,0.780392169952,0.913725495338,0.752941191196});
-    out.push_back(std::vector<double>{0.375,0.631372570992,0.850980401039,0.607843160629});
-    out.push_back(std::vector<double>{0.5,0.454901963472,0.768627464771,0.46274510026});
-    out.push_back(std::vector<double>{0.625,0.254901975393,0.670588254929,0.364705890417});
-    out.push_back(std::vector<double>{0.75,0.137254908681,0.54509806633,0.270588248968});
-    out.push_back(std::vector<double>{0.875,0.0,0.427450984716,0.172549024224});
-    out.push_back(std::vector<double>{1.0,0.0,0.266666680574,0.105882354081});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Greys(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,1.0,1.0});
-    out.push_back(std::vector<double>{0.125,0.941176474094,0.941176474094,0.941176474094});
-    out.push_back(std::vector<double>{0.25,0.850980401039,0.850980401039,0.850980401039});
-    out.push_back(std::vector<double>{0.375,0.741176486015,0.741176486015,0.741176486015});
-    out.push_back(std::vector<double>{0.5,0.588235318661,0.588235318661,0.588235318661});
-    out.push_back(std::vector<double>{0.625,0.450980395079,0.450980395079,0.450980395079});
-    out.push_back(std::vector<double>{0.75,0.321568638086,0.321568638086,0.321568638086});
-    out.push_back(std::vector<double>{0.875,0.145098045468,0.145098045468,0.145098045468});
-    out.push_back(std::vector<double>{1.0,0.0,0.0,0.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Oranges(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,0.960784316063,0.921568632126});
-    out.push_back(std::vector<double>{0.125,0.996078431606,0.901960790157,0.807843148708});
-    out.push_back(std::vector<double>{0.25,0.992156863213,0.815686285496,0.635294139385});
-    out.push_back(std::vector<double>{0.375,0.992156863213,0.68235296011,0.419607847929});
-    out.push_back(std::vector<double>{0.5,0.992156863213,0.552941203117,0.235294118524});
-    out.push_back(std::vector<double>{0.625,0.945098042488,0.411764711142,0.074509806931});
-    out.push_back(std::vector<double>{0.75,0.850980401039,0.282352954149,0.00392156885937});
-    out.push_back(std::vector<double>{0.875,0.65098041296,0.211764708161,0.0117647061124});
-    out.push_back(std::vector<double>{1.0,0.498039215803,0.152941182256,0.0156862754375});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_OrRd(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,0.96862745285,0.92549020052});
-    out.push_back(std::vector<double>{0.125,0.996078431606,0.909803926945,0.784313738346});
-    out.push_back(std::vector<double>{0.25,0.992156863213,0.831372559071,0.61960786581});
-    out.push_back(std::vector<double>{0.375,0.992156863213,0.733333349228,0.517647087574});
-    out.push_back(std::vector<double>{0.5,0.988235294819,0.552941203117,0.349019616842});
-    out.push_back(std::vector<double>{0.625,0.937254905701,0.396078437567,0.282352954149});
-    out.push_back(std::vector<double>{0.75,0.843137264252,0.188235297799,0.121568627656});
-    out.push_back(std::vector<double>{0.875,0.701960802078,0.0,0.0});
-    out.push_back(std::vector<double>{1.0,0.498039215803,0.0,0.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_PuBu(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,0.96862745285,0.984313726425});
-    out.push_back(std::vector<double>{0.125,0.92549020052,0.905882358551,0.949019610882});
-    out.push_back(std::vector<double>{0.25,0.815686285496,0.819607853889,0.901960790157});
-    out.push_back(std::vector<double>{0.375,0.65098041296,0.741176486015,0.858823537827});
-    out.push_back(std::vector<double>{0.5,0.454901963472,0.662745118141,0.811764717102});
-    out.push_back(std::vector<double>{0.625,0.211764708161,0.564705908298,0.752941191196});
-    out.push_back(std::vector<double>{0.75,0.0196078438312,0.439215689898,0.690196096897});
-    out.push_back(std::vector<double>{0.875,0.0156862754375,0.352941185236,0.552941203117});
-    out.push_back(std::vector<double>{1.0,0.00784313771874,0.219607844949,0.345098048449});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_PuBuGn(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,0.96862745285,0.984313726425});
-    out.push_back(std::vector<double>{0.125,0.92549020052,0.886274516582,0.941176474094});
-    out.push_back(std::vector<double>{0.25,0.815686285496,0.819607853889,0.901960790157});
-    out.push_back(std::vector<double>{0.375,0.65098041296,0.741176486015,0.858823537827});
-    out.push_back(std::vector<double>{0.5,0.403921574354,0.662745118141,0.811764717102});
-    out.push_back(std::vector<double>{0.625,0.211764708161,0.564705908298,0.752941191196});
-    out.push_back(std::vector<double>{0.75,0.00784313771874,0.505882382393,0.541176497936});
-    out.push_back(std::vector<double>{0.875,0.00392156885937,0.423529416323,0.349019616842});
-    out.push_back(std::vector<double>{1.0,0.00392156885937,0.274509817362,0.211764708161});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_PuRd(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.96862745285,0.956862747669,0.976470589638});
-    out.push_back(std::vector<double>{0.125,0.905882358551,0.882352948189,0.937254905701});
-    out.push_back(std::vector<double>{0.25,0.831372559071,0.72549021244,0.854901969433});
-    out.push_back(std::vector<double>{0.375,0.78823530674,0.580392181873,0.780392169952});
-    out.push_back(std::vector<double>{0.5,0.874509811401,0.396078437567,0.690196096897});
-    out.push_back(std::vector<double>{0.625,0.905882358551,0.160784319043,0.541176497936});
-    out.push_back(std::vector<double>{0.75,0.807843148708,0.0705882385373,0.337254911661});
-    out.push_back(std::vector<double>{0.875,0.596078455448,0.0,0.262745112181});
-    out.push_back(std::vector<double>{1.0,0.403921574354,0.0,0.121568627656});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Purples(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.988235294819,0.984313726425,0.992156863213});
-    out.push_back(std::vector<double>{0.125,0.937254905701,0.929411768913,0.960784316063});
-    out.push_back(std::vector<double>{0.25,0.854901969433,0.854901969433,0.921568632126});
-    out.push_back(std::vector<double>{0.375,0.737254917622,0.741176486015,0.86274510622});
-    out.push_back(std::vector<double>{0.5,0.61960786581,0.603921592236,0.784313738346});
-    out.push_back(std::vector<double>{0.625,0.501960813999,0.490196079016,0.729411780834});
-    out.push_back(std::vector<double>{0.75,0.415686279535,0.317647069693,0.639215707779});
-    out.push_back(std::vector<double>{0.875,0.329411774874,0.152941182256,0.560784339905});
-    out.push_back(std::vector<double>{1.0,0.247058823705,0.0,0.490196079016});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_RdPu(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,0.96862745285,0.952941179276});
-    out.push_back(std::vector<double>{0.125,0.992156863213,0.878431379795,0.866666674614});
-    out.push_back(std::vector<double>{0.25,0.988235294819,0.772549033165,0.752941191196});
-    out.push_back(std::vector<double>{0.375,0.980392158031,0.623529434204,0.709803938866});
-    out.push_back(std::vector<double>{0.5,0.96862745285,0.407843142748,0.631372570992});
-    out.push_back(std::vector<double>{0.625,0.866666674614,0.203921571374,0.592156887054});
-    out.push_back(std::vector<double>{0.75,0.68235296011,0.00392156885937,0.494117647409});
-    out.push_back(std::vector<double>{0.875,0.478431373835,0.00392156885937,0.466666668653});
-    out.push_back(std::vector<double>{1.0,0.286274522543,0.0,0.415686279535});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Reds(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,0.960784316063,0.941176474094});
-    out.push_back(std::vector<double>{0.125,0.996078431606,0.878431379795,0.823529422283});
-    out.push_back(std::vector<double>{0.25,0.988235294819,0.733333349228,0.631372570992});
-    out.push_back(std::vector<double>{0.375,0.988235294819,0.572549045086,0.447058826685});
-    out.push_back(std::vector<double>{0.5,0.984313726425,0.415686279535,0.290196090937});
-    out.push_back(std::vector<double>{0.625,0.937254905701,0.23137255013,0.172549024224});
-    out.push_back(std::vector<double>{0.75,0.796078443527,0.0941176488996,0.113725490868});
-    out.push_back(std::vector<double>{0.875,0.647058844566,0.0588235296309,0.0823529437184});
-    out.push_back(std::vector<double>{1.0,0.403921574354,0.0,0.0509803928435});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_YlGn(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,1.0,0.898039221764});
-    out.push_back(std::vector<double>{0.125,0.96862745285,0.988235294819,0.72549021244});
-    out.push_back(std::vector<double>{0.25,0.850980401039,0.941176474094,0.639215707779});
-    out.push_back(std::vector<double>{0.375,0.678431391716,0.866666674614,0.556862771511});
-    out.push_back(std::vector<double>{0.5,0.470588237047,0.776470601559,0.474509805441});
-    out.push_back(std::vector<double>{0.625,0.254901975393,0.670588254929,0.364705890417});
-    out.push_back(std::vector<double>{0.75,0.137254908681,0.517647087574,0.262745112181});
-    out.push_back(std::vector<double>{0.875,0.0,0.407843142748,0.215686276555});
-    out.push_back(std::vector<double>{1.0,0.0,0.270588248968,0.160784319043});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_YlGnBu(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,1.0,0.850980401039});
-    out.push_back(std::vector<double>{0.125,0.929411768913,0.972549021244,0.694117665291});
-    out.push_back(std::vector<double>{0.25,0.780392169952,0.913725495338,0.705882370472});
-    out.push_back(std::vector<double>{0.375,0.498039215803,0.803921580315,0.733333349228});
-    out.push_back(std::vector<double>{0.5,0.254901975393,0.713725507259,0.768627464771});
-    out.push_back(std::vector<double>{0.625,0.113725490868,0.568627476692,0.752941191196});
-    out.push_back(std::vector<double>{0.75,0.133333340287,0.368627458811,0.658823549747});
-    out.push_back(std::vector<double>{0.875,0.145098045468,0.203921571374,0.580392181873});
-    out.push_back(std::vector<double>{1.0,0.0313725508749,0.113725490868,0.345098048449});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_YlOrBr(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,1.0,0.898039221764});
-    out.push_back(std::vector<double>{0.125,1.0,0.96862745285,0.737254917622});
-    out.push_back(std::vector<double>{0.25,0.996078431606,0.890196084976,0.568627476692});
-    out.push_back(std::vector<double>{0.375,0.996078431606,0.768627464771,0.309803932905});
-    out.push_back(std::vector<double>{0.5,0.996078431606,0.600000023842,0.160784319043});
-    out.push_back(std::vector<double>{0.625,0.92549020052,0.439215689898,0.0784313753247});
-    out.push_back(std::vector<double>{0.75,0.800000011921,0.298039227724,0.00784313771874});
-    out.push_back(std::vector<double>{0.875,0.600000023842,0.203921571374,0.0156862754375});
-    out.push_back(std::vector<double>{1.0,0.40000000596,0.145098045468,0.0235294122249});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_YlOrRd(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,1.0,0.800000011921});
-    out.push_back(std::vector<double>{0.125,1.0,0.929411768913,0.627451002598});
-    out.push_back(std::vector<double>{0.25,0.996078431606,0.850980401039,0.46274510026});
-    out.push_back(std::vector<double>{0.375,0.996078431606,0.698039233685,0.298039227724});
-    out.push_back(std::vector<double>{0.5,0.992156863213,0.552941203117,0.235294118524});
-    out.push_back(std::vector<double>{0.625,0.988235294819,0.305882364511,0.164705887437});
-    out.push_back(std::vector<double>{0.75,0.890196084976,0.101960785687,0.109803922474});
-    out.push_back(std::vector<double>{0.875,0.741176486015,0.0,0.149019613862});
-    out.push_back(std::vector<double>{1.0,0.501960813999,0.0,0.149019613862});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_afmhot(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,-0.5,-1.0});
-    out.push_back(std::vector<double>{0.01,0.02,-0.48,-0.98});
-    out.push_back(std::vector<double>{0.02,0.04,-0.46,-0.96});
-    out.push_back(std::vector<double>{0.03,0.06,-0.44,-0.94});
-    out.push_back(std::vector<double>{0.04,0.08,-0.42,-0.92});
-    out.push_back(std::vector<double>{0.05,0.1,-0.4,-0.9});
-    out.push_back(std::vector<double>{0.06,0.12,-0.38,-0.88});
-    out.push_back(std::vector<double>{0.07,0.14,-0.36,-0.86});
-    out.push_back(std::vector<double>{0.08,0.16,-0.34,-0.84});
-    out.push_back(std::vector<double>{0.09,0.18,-0.32,-0.82});
-    out.push_back(std::vector<double>{0.1,0.2,-0.3,-0.8});
-    out.push_back(std::vector<double>{0.11,0.22,-0.28,-0.78});
-    out.push_back(std::vector<double>{0.12,0.24,-0.26,-0.76});
-    out.push_back(std::vector<double>{0.13,0.26,-0.24,-0.74});
-    out.push_back(std::vector<double>{0.14,0.28,-0.22,-0.72});
-    out.push_back(std::vector<double>{0.15,0.3,-0.2,-0.7});
-    out.push_back(std::vector<double>{0.16,0.32,-0.18,-0.68});
-    out.push_back(std::vector<double>{0.17,0.34,-0.16,-0.66});
-    out.push_back(std::vector<double>{0.18,0.36,-0.14,-0.64});
-    out.push_back(std::vector<double>{0.19,0.38,-0.12,-0.62});
-    out.push_back(std::vector<double>{0.2,0.4,-0.1,-0.6});
-    out.push_back(std::vector<double>{0.21,0.42,-0.08,-0.58});
-    out.push_back(std::vector<double>{0.22,0.44,-0.06,-0.56});
-    out.push_back(std::vector<double>{0.23,0.46,-0.04,-0.54});
-    out.push_back(std::vector<double>{0.24,0.48,-0.02,-0.52});
-    out.push_back(std::vector<double>{0.25,0.5,0.0,-0.5});
-    out.push_back(std::vector<double>{0.26,0.52,0.02,-0.48});
-    out.push_back(std::vector<double>{0.27,0.54,0.04,-0.46});
-    out.push_back(std::vector<double>{0.28,0.56,0.06,-0.44});
-    out.push_back(std::vector<double>{0.29,0.58,0.08,-0.42});
-    out.push_back(std::vector<double>{0.3,0.6,0.1,-0.4});
-    out.push_back(std::vector<double>{0.31,0.62,0.12,-0.38});
-    out.push_back(std::vector<double>{0.32,0.64,0.14,-0.36});
-    out.push_back(std::vector<double>{0.33,0.66,0.16,-0.34});
-    out.push_back(std::vector<double>{0.34,0.68,0.18,-0.32});
-    out.push_back(std::vector<double>{0.35,0.7,0.2,-0.3});
-    out.push_back(std::vector<double>{0.36,0.72,0.22,-0.28});
-    out.push_back(std::vector<double>{0.37,0.74,0.24,-0.26});
-    out.push_back(std::vector<double>{0.38,0.76,0.26,-0.24});
-    out.push_back(std::vector<double>{0.39,0.78,0.28,-0.22});
-    out.push_back(std::vector<double>{0.4,0.8,0.3,-0.2});
-    out.push_back(std::vector<double>{0.41,0.82,0.32,-0.18});
-    out.push_back(std::vector<double>{0.42,0.84,0.34,-0.16});
-    out.push_back(std::vector<double>{0.43,0.86,0.36,-0.14});
-    out.push_back(std::vector<double>{0.44,0.88,0.38,-0.12});
-    out.push_back(std::vector<double>{0.45,0.9,0.4,-0.1});
-    out.push_back(std::vector<double>{0.46,0.92,0.42,-0.08});
-    out.push_back(std::vector<double>{0.47,0.94,0.44,-0.06});
-    out.push_back(std::vector<double>{0.48,0.96,0.46,-0.04});
-    out.push_back(std::vector<double>{0.49,0.98,0.48,-0.02});
-    out.push_back(std::vector<double>{0.5,1.0,0.5,0.0});
-    out.push_back(std::vector<double>{0.51,1.02,0.52,0.02});
-    out.push_back(std::vector<double>{0.52,1.04,0.54,0.04});
-    out.push_back(std::vector<double>{0.53,1.06,0.56,0.06});
-    out.push_back(std::vector<double>{0.54,1.08,0.58,0.08});
-    out.push_back(std::vector<double>{0.55,1.1,0.6,0.1});
-    out.push_back(std::vector<double>{0.56,1.12,0.62,0.12});
-    out.push_back(std::vector<double>{0.57,1.14,0.64,0.14});
-    out.push_back(std::vector<double>{0.58,1.16,0.66,0.16});
-    out.push_back(std::vector<double>{0.59,1.18,0.68,0.18});
-    out.push_back(std::vector<double>{0.6,1.2,0.7,0.2});
-    out.push_back(std::vector<double>{0.61,1.22,0.72,0.22});
-    out.push_back(std::vector<double>{0.62,1.24,0.74,0.24});
-    out.push_back(std::vector<double>{0.63,1.26,0.76,0.26});
-    out.push_back(std::vector<double>{0.64,1.28,0.78,0.28});
-    out.push_back(std::vector<double>{0.65,1.3,0.8,0.3});
-    out.push_back(std::vector<double>{0.66,1.32,0.82,0.32});
-    out.push_back(std::vector<double>{0.67,1.34,0.84,0.34});
-    out.push_back(std::vector<double>{0.68,1.36,0.86,0.36});
-    out.push_back(std::vector<double>{0.69,1.38,0.88,0.38});
-    out.push_back(std::vector<double>{0.7,1.4,0.9,0.4});
-    out.push_back(std::vector<double>{0.71,1.42,0.92,0.42});
-    out.push_back(std::vector<double>{0.72,1.44,0.94,0.44});
-    out.push_back(std::vector<double>{0.73,1.46,0.96,0.46});
-    out.push_back(std::vector<double>{0.74,1.48,0.98,0.48});
-    out.push_back(std::vector<double>{0.75,1.5,1.0,0.5});
-    out.push_back(std::vector<double>{0.76,1.52,1.02,0.52});
-    out.push_back(std::vector<double>{0.77,1.54,1.04,0.54});
-    out.push_back(std::vector<double>{0.78,1.56,1.06,0.56});
-    out.push_back(std::vector<double>{0.79,1.58,1.08,0.58});
-    out.push_back(std::vector<double>{0.8,1.6,1.1,0.6});
-    out.push_back(std::vector<double>{0.81,1.62,1.12,0.62});
-    out.push_back(std::vector<double>{0.82,1.64,1.14,0.64});
-    out.push_back(std::vector<double>{0.83,1.66,1.16,0.66});
-    out.push_back(std::vector<double>{0.84,1.68,1.18,0.68});
-    out.push_back(std::vector<double>{0.85,1.7,1.2,0.7});
-    out.push_back(std::vector<double>{0.86,1.72,1.22,0.72});
-    out.push_back(std::vector<double>{0.87,1.74,1.24,0.74});
-    out.push_back(std::vector<double>{0.88,1.76,1.26,0.76});
-    out.push_back(std::vector<double>{0.89,1.78,1.28,0.78});
-    out.push_back(std::vector<double>{0.9,1.8,1.3,0.8});
-    out.push_back(std::vector<double>{0.91,1.82,1.32,0.82});
-    out.push_back(std::vector<double>{0.92,1.84,1.34,0.84});
-    out.push_back(std::vector<double>{0.93,1.86,1.36,0.86});
-    out.push_back(std::vector<double>{0.94,1.88,1.38,0.88});
-    out.push_back(std::vector<double>{0.95,1.9,1.4,0.9});
-    out.push_back(std::vector<double>{0.96,1.92,1.42,0.92});
-    out.push_back(std::vector<double>{0.97,1.94,1.44,0.94});
-    out.push_back(std::vector<double>{0.98,1.96,1.46,0.96});
-    out.push_back(std::vector<double>{0.99,1.98,1.48,0.98});
-    out.push_back(std::vector<double>{1.0,2.0,1.5,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_autumn(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,0.0,0.0});
-    out.push_back(std::vector<double>{1.0,1.0,1.0,0.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_bone(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,0.0});
-    out.push_back(std::vector<double>{0.365079,0.319444125,0.319444,0.444444});
-    out.push_back(std::vector<double>{0.746032,0.652778,0.777778,0.77777795});
-    out.push_back(std::vector<double>{1.0,1.0,1.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_cool(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,1.0,1.0});
-    out.push_back(std::vector<double>{1.0,1.0,0.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_copper(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,0.0});
-    out.push_back(std::vector<double>{0.809524,1.0,0.6324001488,0.40273819});
-    out.push_back(std::vector<double>{1.0,1.0,0.7812,0.4975});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_gist_heat(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,-1.0,-3.0});
-    out.push_back(std::vector<double>{0.01,0.015,-0.98,-2.96});
-    out.push_back(std::vector<double>{0.02,0.03,-0.96,-2.92});
-    out.push_back(std::vector<double>{0.03,0.045,-0.94,-2.88});
-    out.push_back(std::vector<double>{0.04,0.06,-0.92,-2.84});
-    out.push_back(std::vector<double>{0.05,0.075,-0.9,-2.8});
-    out.push_back(std::vector<double>{0.06,0.09,-0.88,-2.76});
-    out.push_back(std::vector<double>{0.07,0.105,-0.86,-2.72});
-    out.push_back(std::vector<double>{0.08,0.12,-0.84,-2.68});
-    out.push_back(std::vector<double>{0.09,0.135,-0.82,-2.64});
-    out.push_back(std::vector<double>{0.1,0.15,-0.8,-2.6});
-    out.push_back(std::vector<double>{0.11,0.165,-0.78,-2.56});
-    out.push_back(std::vector<double>{0.12,0.18,-0.76,-2.52});
-    out.push_back(std::vector<double>{0.13,0.195,-0.74,-2.48});
-    out.push_back(std::vector<double>{0.14,0.21,-0.72,-2.44});
-    out.push_back(std::vector<double>{0.15,0.225,-0.7,-2.4});
-    out.push_back(std::vector<double>{0.16,0.24,-0.68,-2.36});
-    out.push_back(std::vector<double>{0.17,0.255,-0.66,-2.32});
-    out.push_back(std::vector<double>{0.18,0.27,-0.64,-2.28});
-    out.push_back(std::vector<double>{0.19,0.285,-0.62,-2.24});
-    out.push_back(std::vector<double>{0.2,0.3,-0.6,-2.2});
-    out.push_back(std::vector<double>{0.21,0.315,-0.58,-2.16});
-    out.push_back(std::vector<double>{0.22,0.33,-0.56,-2.12});
-    out.push_back(std::vector<double>{0.23,0.345,-0.54,-2.08});
-    out.push_back(std::vector<double>{0.24,0.36,-0.52,-2.04});
-    out.push_back(std::vector<double>{0.25,0.375,-0.5,-2.0});
-    out.push_back(std::vector<double>{0.26,0.39,-0.48,-1.96});
-    out.push_back(std::vector<double>{0.27,0.405,-0.46,-1.92});
-    out.push_back(std::vector<double>{0.28,0.42,-0.44,-1.88});
-    out.push_back(std::vector<double>{0.29,0.435,-0.42,-1.84});
-    out.push_back(std::vector<double>{0.3,0.45,-0.4,-1.8});
-    out.push_back(std::vector<double>{0.31,0.465,-0.38,-1.76});
-    out.push_back(std::vector<double>{0.32,0.48,-0.36,-1.72});
-    out.push_back(std::vector<double>{0.33,0.495,-0.34,-1.68});
-    out.push_back(std::vector<double>{0.34,0.51,-0.32,-1.64});
-    out.push_back(std::vector<double>{0.35,0.525,-0.3,-1.6});
-    out.push_back(std::vector<double>{0.36,0.54,-0.28,-1.56});
-    out.push_back(std::vector<double>{0.37,0.555,-0.26,-1.52});
-    out.push_back(std::vector<double>{0.38,0.57,-0.24,-1.48});
-    out.push_back(std::vector<double>{0.39,0.585,-0.22,-1.44});
-    out.push_back(std::vector<double>{0.4,0.6,-0.2,-1.4});
-    out.push_back(std::vector<double>{0.41,0.615,-0.18,-1.36});
-    out.push_back(std::vector<double>{0.42,0.63,-0.16,-1.32});
-    out.push_back(std::vector<double>{0.43,0.645,-0.14,-1.28});
-    out.push_back(std::vector<double>{0.44,0.66,-0.12,-1.24});
-    out.push_back(std::vector<double>{0.45,0.675,-0.1,-1.2});
-    out.push_back(std::vector<double>{0.46,0.69,-0.08,-1.16});
-    out.push_back(std::vector<double>{0.47,0.705,-0.06,-1.12});
-    out.push_back(std::vector<double>{0.48,0.72,-0.04,-1.08});
-    out.push_back(std::vector<double>{0.49,0.735,-0.02,-1.04});
-    out.push_back(std::vector<double>{0.5,0.75,0.0,-1.0});
-    out.push_back(std::vector<double>{0.51,0.765,0.02,-0.96});
-    out.push_back(std::vector<double>{0.52,0.78,0.04,-0.92});
-    out.push_back(std::vector<double>{0.53,0.795,0.06,-0.88});
-    out.push_back(std::vector<double>{0.54,0.81,0.08,-0.84});
-    out.push_back(std::vector<double>{0.55,0.825,0.1,-0.8});
-    out.push_back(std::vector<double>{0.56,0.84,0.12,-0.76});
-    out.push_back(std::vector<double>{0.57,0.855,0.14,-0.72});
-    out.push_back(std::vector<double>{0.58,0.87,0.16,-0.68});
-    out.push_back(std::vector<double>{0.59,0.885,0.18,-0.64});
-    out.push_back(std::vector<double>{0.6,0.9,0.2,-0.6});
-    out.push_back(std::vector<double>{0.61,0.915,0.22,-0.56});
-    out.push_back(std::vector<double>{0.62,0.93,0.24,-0.52});
-    out.push_back(std::vector<double>{0.63,0.945,0.26,-0.48});
-    out.push_back(std::vector<double>{0.64,0.96,0.28,-0.44});
-    out.push_back(std::vector<double>{0.65,0.975,0.3,-0.4});
-    out.push_back(std::vector<double>{0.66,0.99,0.32,-0.36});
-    out.push_back(std::vector<double>{0.67,1.005,0.34,-0.32});
-    out.push_back(std::vector<double>{0.68,1.02,0.36,-0.28});
-    out.push_back(std::vector<double>{0.69,1.035,0.38,-0.24});
-    out.push_back(std::vector<double>{0.7,1.05,0.4,-0.2});
-    out.push_back(std::vector<double>{0.71,1.065,0.42,-0.16});
-    out.push_back(std::vector<double>{0.72,1.08,0.44,-0.12});
-    out.push_back(std::vector<double>{0.73,1.095,0.46,-0.08});
-    out.push_back(std::vector<double>{0.74,1.11,0.48,-0.04});
-    out.push_back(std::vector<double>{0.75,1.125,0.5,0.0});
-    out.push_back(std::vector<double>{0.76,1.14,0.52,0.04});
-    out.push_back(std::vector<double>{0.77,1.155,0.54,0.08});
-    out.push_back(std::vector<double>{0.78,1.17,0.56,0.12});
-    out.push_back(std::vector<double>{0.79,1.185,0.58,0.16});
-    out.push_back(std::vector<double>{0.8,1.2,0.6,0.2});
-    out.push_back(std::vector<double>{0.81,1.215,0.62,0.24});
-    out.push_back(std::vector<double>{0.82,1.23,0.64,0.28});
-    out.push_back(std::vector<double>{0.83,1.245,0.66,0.32});
-    out.push_back(std::vector<double>{0.84,1.26,0.68,0.36});
-    out.push_back(std::vector<double>{0.85,1.275,0.7,0.4});
-    out.push_back(std::vector<double>{0.86,1.29,0.72,0.44});
-    out.push_back(std::vector<double>{0.87,1.305,0.74,0.48});
-    out.push_back(std::vector<double>{0.88,1.32,0.76,0.52});
-    out.push_back(std::vector<double>{0.89,1.335,0.78,0.56});
-    out.push_back(std::vector<double>{0.9,1.35,0.8,0.6});
-    out.push_back(std::vector<double>{0.91,1.365,0.82,0.64});
-    out.push_back(std::vector<double>{0.92,1.38,0.84,0.68});
-    out.push_back(std::vector<double>{0.93,1.395,0.86,0.72});
-    out.push_back(std::vector<double>{0.94,1.41,0.88,0.76});
-    out.push_back(std::vector<double>{0.95,1.425,0.9,0.8});
-    out.push_back(std::vector<double>{0.96,1.44,0.92,0.84});
-    out.push_back(std::vector<double>{0.97,1.455,0.94,0.88});
-    out.push_back(std::vector<double>{0.98,1.47,0.96,0.92});
-    out.push_back(std::vector<double>{0.99,1.485,0.98,0.96});
-    out.push_back(std::vector<double>{1.0,1.5,1.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_gray(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0,0,0});
-    out.push_back(std::vector<double>{1.0,1,1,1});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_hot(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0416,0.0,0.0});
-    out.push_back(std::vector<double>{0.365079,1.0,0.0,0.0});
-    out.push_back(std::vector<double>{0.746032,1.0,1.0,0.0});
-    out.push_back(std::vector<double>{1.0,1.0,1.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_pink(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.1178,0.0,0.0});
-    out.push_back(std::vector<double>{0.015873,0.195857,0.102869,0.102869});
-    out.push_back(std::vector<double>{0.031746,0.250661,0.145479,0.145479});
-    out.push_back(std::vector<double>{0.047619,0.295468,0.178174,0.178174});
-    out.push_back(std::vector<double>{0.063492,0.334324,0.205738,0.205738});
-    out.push_back(std::vector<double>{0.079365,0.369112,0.230022,0.230022});
-    out.push_back(std::vector<double>{0.095238,0.400892,0.251976,0.251976});
-    out.push_back(std::vector<double>{0.111111,0.430331,0.272166,0.272166});
-    out.push_back(std::vector<double>{0.126984,0.457882,0.290957,0.290957});
-    out.push_back(std::vector<double>{0.142857,0.483867,0.308607,0.308607});
-    out.push_back(std::vector<double>{0.15873,0.508525,0.3253,0.3253});
-    out.push_back(std::vector<double>{0.174603,0.532042,0.341178,0.341178});
-    out.push_back(std::vector<double>{0.190476,0.554563,0.356348,0.356348});
-    out.push_back(std::vector<double>{0.206349,0.576204,0.370899,0.370899});
-    out.push_back(std::vector<double>{0.222222,0.597061,0.3849,0.3849});
-    out.push_back(std::vector<double>{0.238095,0.617213,0.39841,0.39841});
-    out.push_back(std::vector<double>{0.253968,0.636729,0.411476,0.411476});
-    out.push_back(std::vector<double>{0.269841,0.655663,0.424139,0.424139});
-    out.push_back(std::vector<double>{0.285714,0.674066,0.436436,0.436436});
-    out.push_back(std::vector<double>{0.301587,0.69198,0.448395,0.448395});
-    out.push_back(std::vector<double>{0.31746,0.709441,0.460044,0.460044});
-    out.push_back(std::vector<double>{0.333333,0.726483,0.471405,0.471405});
-    out.push_back(std::vector<double>{0.349206,0.743134,0.482498,0.482498});
-    out.push_back(std::vector<double>{0.365079,0.759421,0.493342,0.493342});
-    out.push_back(std::vector<double>{0.380952,0.766356,0.517549,0.503953});
-    out.push_back(std::vector<double>{0.396825,0.773229,0.540674,0.514344});
-    out.push_back(std::vector<double>{0.412698,0.780042,0.562849,0.524531});
-    out.push_back(std::vector<double>{0.428571,0.786796,0.584183,0.534522});
-    out.push_back(std::vector<double>{0.444444,0.793492,0.604765,0.544331});
-    out.push_back(std::vector<double>{0.460317,0.800132,0.624669,0.553966});
-    out.push_back(std::vector<double>{0.47619,0.806718,0.643958,0.563436});
-    out.push_back(std::vector<double>{0.492063,0.81325,0.662687,0.57275});
-    out.push_back(std::vector<double>{0.507937,0.81973,0.6809,0.581914});
-    out.push_back(std::vector<double>{0.52381,0.82616,0.698638,0.590937});
-    out.push_back(std::vector<double>{0.539683,0.832539,0.715937,0.599824});
-    out.push_back(std::vector<double>{0.555556,0.83887,0.732828,0.608581});
-    out.push_back(std::vector<double>{0.571429,0.845154,0.749338,0.617213});
-    out.push_back(std::vector<double>{0.587302,0.851392,0.765493,0.625727});
-    out.push_back(std::vector<double>{0.603175,0.857584,0.781313,0.634126});
-    out.push_back(std::vector<double>{0.619048,0.863731,0.796819,0.642416});
-    out.push_back(std::vector<double>{0.634921,0.869835,0.812029,0.6506});
-    out.push_back(std::vector<double>{0.650794,0.875897,0.82696,0.658682});
-    out.push_back(std::vector<double>{0.666667,0.881917,0.841625,0.666667});
-    out.push_back(std::vector<double>{0.68254,0.887896,0.85604,0.674556});
-    out.push_back(std::vector<double>{0.698413,0.893835,0.870216,0.682355});
-    out.push_back(std::vector<double>{0.714286,0.899735,0.884164,0.690066});
-    out.push_back(std::vector<double>{0.730159,0.905597,0.897896,0.697691});
-    out.push_back(std::vector<double>{0.746032,0.911421,0.911421,0.705234});
-    out.push_back(std::vector<double>{0.761905,0.917208,0.917208,0.727166});
-    out.push_back(std::vector<double>{0.777778,0.922958,0.922958,0.748455});
-    out.push_back(std::vector<double>{0.793651,0.928673,0.928673,0.769156});
-    out.push_back(std::vector<double>{0.809524,0.934353,0.934353,0.789314});
-    out.push_back(std::vector<double>{0.825397,0.939999,0.939999,0.808969});
-    out.push_back(std::vector<double>{0.84127,0.945611,0.945611,0.828159});
-    out.push_back(std::vector<double>{0.857143,0.95119,0.95119,0.846913});
-    out.push_back(std::vector<double>{0.873016,0.956736,0.956736,0.865261});
-    out.push_back(std::vector<double>{0.888889,0.96225,0.96225,0.883229});
-    out.push_back(std::vector<double>{0.904762,0.967733,0.967733,0.900837});
-    out.push_back(std::vector<double>{0.920635,0.973185,0.973185,0.918109});
-    out.push_back(std::vector<double>{0.936508,0.978607,0.978607,0.935061});
-    out.push_back(std::vector<double>{0.952381,0.983999,0.983999,0.951711});
-    out.push_back(std::vector<double>{0.968254,0.989361,0.989361,0.968075});
-    out.push_back(std::vector<double>{0.984127,0.994695,0.994695,0.984167});
-    out.push_back(std::vector<double>{1.0,1.0,1.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_spring(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,0.0,1.0});
-    out.push_back(std::vector<double>{1.0,1.0,1.0,0.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_summer(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.5,0.4});
-    out.push_back(std::vector<double>{1.0,1.0,1.0,0.4});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_winter(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,1.0});
-    out.push_back(std::vector<double>{1.0,0.0,1.0,0.5});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_BrBG(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.329411774874,0.188235297799,0.0196078438312});
-    out.push_back(std::vector<double>{0.1,0.549019634724,0.317647069693,0.0392156876624});
-    out.push_back(std::vector<double>{0.2,0.749019622803,0.505882382393,0.176470592618});
-    out.push_back(std::vector<double>{0.3,0.874509811401,0.760784327984,0.490196079016});
-    out.push_back(std::vector<double>{0.4,0.964705884457,0.909803926945,0.764705896378});
-    out.push_back(std::vector<double>{0.5,0.960784316063,0.960784316063,0.960784316063});
-    out.push_back(std::vector<double>{0.6,0.780392169952,0.917647063732,0.898039221764});
-    out.push_back(std::vector<double>{0.7,0.501960813999,0.803921580315,0.75686275959});
-    out.push_back(std::vector<double>{0.8,0.207843139768,0.592156887054,0.560784339905});
-    out.push_back(std::vector<double>{0.9,0.00392156885937,0.40000000596,0.368627458811});
-    out.push_back(std::vector<double>{1.0,0.0,0.235294118524,0.188235297799});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_bwr(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,1.0,1.0});
-    out.push_back(std::vector<double>{0.5,1.0,1.0,1.0,1.0});
-    out.push_back(std::vector<double>{1.0,1.0,0.0,0.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_coolwarm(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.2298057,0.298717966,0.753683153});
-    out.push_back(std::vector<double>{0.03125,0.26623388,0.353094838,0.801466763});
-    out.push_back(std::vector<double>{0.0625,0.30386891,0.406535296,0.84495867});
-    out.push_back(std::vector<double>{0.09375,0.342804478,0.458757618,0.883725899});
-    out.push_back(std::vector<double>{0.125,0.38301334,0.50941904,0.917387822});
-    out.push_back(std::vector<double>{0.15625,0.424369608,0.558148092,0.945619588});
-    out.push_back(std::vector<double>{0.1875,0.46666708,0.604562568,0.968154911});
-    out.push_back(std::vector<double>{0.21875,0.509635204,0.648280772,0.98478814});
-    out.push_back(std::vector<double>{0.25,0.552953156,0.688929332,0.995375608});
-    out.push_back(std::vector<double>{0.28125,0.596262162,0.726149107,0.999836203});
-    out.push_back(std::vector<double>{0.3125,0.639176211,0.759599947,0.998151185});
-    out.push_back(std::vector<double>{0.34375,0.681291281,0.788964712,0.990363227});
-    out.push_back(std::vector<double>{0.375,0.722193294,0.813952739,0.976574709});
-    out.push_back(std::vector<double>{0.40625,0.761464949,0.834302879,0.956945269});
-    out.push_back(std::vector<double>{0.4375,0.798691636,0.849786142,0.931688648});
-    out.push_back(std::vector<double>{0.46875,0.833466556,0.860207984,0.901068838});
-    out.push_back(std::vector<double>{0.5,0.865395197,0.86541021,0.865395561});
-    out.push_back(std::vector<double>{0.53125,0.897787179,0.848937047,0.820880546});
-    out.push_back(std::vector<double>{0.5625,0.924127593,0.827384882,0.774508472});
-    out.push_back(std::vector<double>{0.59375,0.944468518,0.800927443,0.726736146});
-    out.push_back(std::vector<double>{0.625,0.958852946,0.769767752,0.678007945});
-    out.push_back(std::vector<double>{0.65625,0.96732803,0.734132809,0.628751763});
-    out.push_back(std::vector<double>{0.6875,0.969954137,0.694266682,0.579375448});
-    out.push_back(std::vector<double>{0.71875,0.966811177,0.650421156,0.530263762});
-    out.push_back(std::vector<double>{0.75,0.958003065,0.602842431,0.481775914});
-    out.push_back(std::vector<double>{0.78125,0.943660866,0.551750968,0.434243684});
-    out.push_back(std::vector<double>{0.8125,0.923944917,0.49730856,0.387970225});
-    out.push_back(std::vector<double>{0.84375,0.89904617,0.439559467,0.343229596});
-    out.push_back(std::vector<double>{0.875,0.869186849,0.378313092,0.300267182});
-    out.push_back(std::vector<double>{0.90625,0.834620542,0.312874446,0.259301199});
-    out.push_back(std::vector<double>{0.9375,0.795631745,0.24128379,0.220525627});
-    out.push_back(std::vector<double>{0.96875,0.752534934,0.157246067,0.184115123});
-    out.push_back(std::vector<double>{1.0,0.705673158,0.01555616,0.150232812});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_PiYG(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.556862771511,0.00392156885937,0.321568638086});
-    out.push_back(std::vector<double>{0.1,0.772549033165,0.105882354081,0.490196079016});
-    out.push_back(std::vector<double>{0.2,0.870588243008,0.466666668653,0.68235296011});
-    out.push_back(std::vector<double>{0.3,0.945098042488,0.713725507259,0.854901969433});
-    out.push_back(std::vector<double>{0.4,0.992156863213,0.878431379795,0.937254905701});
-    out.push_back(std::vector<double>{0.5,0.96862745285,0.96862745285,0.96862745285});
-    out.push_back(std::vector<double>{0.6,0.901960790157,0.960784316063,0.815686285496});
-    out.push_back(std::vector<double>{0.7,0.721568644047,0.882352948189,0.525490224361});
-    out.push_back(std::vector<double>{0.8,0.498039215803,0.737254917622,0.254901975393});
-    out.push_back(std::vector<double>{0.9,0.301960796118,0.572549045086,0.129411771894});
-    out.push_back(std::vector<double>{1.0,0.152941182256,0.392156869173,0.0980392172933});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_PRGn(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.250980407,0.0,0.29411765933});
-    out.push_back(std::vector<double>{0.1,0.46274510026,0.164705887437,0.51372551918});
-    out.push_back(std::vector<double>{0.2,0.600000023842,0.439215689898,0.670588254929});
-    out.push_back(std::vector<double>{0.3,0.760784327984,0.647058844566,0.811764717102});
-    out.push_back(std::vector<double>{0.4,0.905882358551,0.831372559071,0.909803926945});
-    out.push_back(std::vector<double>{0.5,0.96862745285,0.96862745285,0.96862745285});
-    out.push_back(std::vector<double>{0.6,0.850980401039,0.941176474094,0.827450990677});
-    out.push_back(std::vector<double>{0.7,0.65098041296,0.858823537827,0.627451002598});
-    out.push_back(std::vector<double>{0.8,0.352941185236,0.68235296011,0.380392163992});
-    out.push_back(std::vector<double>{0.9,0.105882354081,0.470588237047,0.215686276555});
-    out.push_back(std::vector<double>{1.0,0.0,0.266666680574,0.105882354081});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_PuOr(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.498039215803,0.23137255013,0.0313725508749});
-    out.push_back(std::vector<double>{0.1,0.701960802078,0.345098048449,0.0235294122249});
-    out.push_back(std::vector<double>{0.2,0.878431379795,0.509803950787,0.0784313753247});
-    out.push_back(std::vector<double>{0.3,0.992156863213,0.721568644047,0.388235300779});
-    out.push_back(std::vector<double>{0.4,0.996078431606,0.878431379795,0.713725507259});
-    out.push_back(std::vector<double>{0.5,0.96862745285,0.96862745285,0.96862745285});
-    out.push_back(std::vector<double>{0.6,0.847058832645,0.854901969433,0.921568632126});
-    out.push_back(std::vector<double>{0.7,0.698039233685,0.670588254929,0.823529422283});
-    out.push_back(std::vector<double>{0.8,0.501960813999,0.450980395079,0.674509823322});
-    out.push_back(std::vector<double>{0.9,0.329411774874,0.152941182256,0.533333361149});
-    out.push_back(std::vector<double>{1.0,0.176470592618,0.0,0.29411765933});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_RdBu(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.403921574354,0.0,0.121568627656});
-    out.push_back(std::vector<double>{0.1,0.698039233685,0.0941176488996,0.168627455831});
-    out.push_back(std::vector<double>{0.2,0.839215695858,0.376470595598,0.301960796118});
-    out.push_back(std::vector<double>{0.3,0.956862747669,0.647058844566,0.509803950787});
-    out.push_back(std::vector<double>{0.4,0.992156863213,0.858823537827,0.780392169952});
-    out.push_back(std::vector<double>{0.5,0.96862745285,0.96862745285,0.96862745285});
-    out.push_back(std::vector<double>{0.6,0.819607853889,0.898039221764,0.941176474094});
-    out.push_back(std::vector<double>{0.7,0.572549045086,0.772549033165,0.870588243008});
-    out.push_back(std::vector<double>{0.8,0.262745112181,0.57647061348,0.764705896378});
-    out.push_back(std::vector<double>{0.9,0.129411771894,0.40000000596,0.674509823322});
-    out.push_back(std::vector<double>{1.0,0.0196078438312,0.188235297799,0.380392163992});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_RdGy(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.403921574354,0.0,0.121568627656});
-    out.push_back(std::vector<double>{0.1,0.698039233685,0.0941176488996,0.168627455831});
-    out.push_back(std::vector<double>{0.2,0.839215695858,0.376470595598,0.301960796118});
-    out.push_back(std::vector<double>{0.3,0.956862747669,0.647058844566,0.509803950787});
-    out.push_back(std::vector<double>{0.4,0.992156863213,0.858823537827,0.780392169952});
-    out.push_back(std::vector<double>{0.5,1.0,1.0,1.0});
-    out.push_back(std::vector<double>{0.6,0.878431379795,0.878431379795,0.878431379795});
-    out.push_back(std::vector<double>{0.7,0.729411780834,0.729411780834,0.729411780834});
-    out.push_back(std::vector<double>{0.8,0.529411792755,0.529411792755,0.529411792755});
-    out.push_back(std::vector<double>{0.9,0.301960796118,0.301960796118,0.301960796118});
-    out.push_back(std::vector<double>{1.0,0.101960785687,0.101960785687,0.101960785687});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_RdYlBu(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.647058844566,0.0,0.149019613862});
-    out.push_back(std::vector<double>{0.10000000149,0.843137264252,0.188235297799,0.152941182256});
-    out.push_back(std::vector<double>{0.20000000298,0.956862747669,0.427450984716,0.262745112181});
-    out.push_back(std::vector<double>{0.300000011921,0.992156863213,0.68235296011,0.380392163992});
-    out.push_back(std::vector<double>{0.40000000596,0.996078431606,0.878431379795,0.564705908298});
-    out.push_back(std::vector<double>{0.5,1.0,1.0,0.749019622803});
-    out.push_back(std::vector<double>{0.600000023842,0.878431379795,0.952941179276,0.972549021244});
-    out.push_back(std::vector<double>{0.699999988079,0.670588254929,0.850980401039,0.913725495338});
-    out.push_back(std::vector<double>{0.800000011921,0.454901963472,0.678431391716,0.819607853889});
-    out.push_back(std::vector<double>{0.899999976158,0.270588248968,0.458823531866,0.705882370472});
-    out.push_back(std::vector<double>{1.0,0.192156866193,0.211764708161,0.584313750267});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_RdYlGn(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.647058844566,0.0,0.149019613862});
-    out.push_back(std::vector<double>{0.1,0.843137264252,0.188235297799,0.152941182256});
-    out.push_back(std::vector<double>{0.2,0.956862747669,0.427450984716,0.262745112181});
-    out.push_back(std::vector<double>{0.3,0.992156863213,0.68235296011,0.380392163992});
-    out.push_back(std::vector<double>{0.4,0.996078431606,0.878431379795,0.54509806633});
-    out.push_back(std::vector<double>{0.5,1.0,1.0,0.749019622803});
-    out.push_back(std::vector<double>{0.6,0.850980401039,0.937254905701,0.54509806633});
-    out.push_back(std::vector<double>{0.7,0.65098041296,0.850980401039,0.415686279535});
-    out.push_back(std::vector<double>{0.8,0.40000000596,0.741176486015,0.388235300779});
-    out.push_back(std::vector<double>{0.9,0.101960785687,0.596078455448,0.313725501299});
-    out.push_back(std::vector<double>{1.0,0.0,0.407843142748,0.215686276555});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Spectral(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.61960786581,0.00392156885937,0.258823543787});
-    out.push_back(std::vector<double>{0.1,0.835294127464,0.243137255311,0.309803932905});
-    out.push_back(std::vector<double>{0.2,0.956862747669,0.427450984716,0.262745112181});
-    out.push_back(std::vector<double>{0.3,0.992156863213,0.68235296011,0.380392163992});
-    out.push_back(std::vector<double>{0.4,0.996078431606,0.878431379795,0.54509806633});
-    out.push_back(std::vector<double>{0.5,1.0,1.0,0.749019622803});
-    out.push_back(std::vector<double>{0.6,0.901960790157,0.960784316063,0.596078455448});
-    out.push_back(std::vector<double>{0.7,0.670588254929,0.866666674614,0.643137276173});
-    out.push_back(std::vector<double>{0.8,0.40000000596,0.760784327984,0.647058844566});
-    out.push_back(std::vector<double>{0.9,0.196078434587,0.533333361149,0.741176486015});
-    out.push_back(std::vector<double>{1.0,0.368627458811,0.309803932905,0.635294139385});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_seismic(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,0.3,1.0});
-    out.push_back(std::vector<double>{0.25,0.0,0.0,1.0,1.0});
-    out.push_back(std::vector<double>{0.5,1.0,1.0,1.0,1.0});
-    out.push_back(std::vector<double>{0.75,1.0,0.0,0.0,1.0});
-    out.push_back(std::vector<double>{1.0,0.5,0.0,0.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Accent(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.498039215803,0.78823530674,0.498039215803});
-    out.push_back(std::vector<double>{0.142857142857,0.745098054409,0.68235296011,0.831372559071});
-    out.push_back(std::vector<double>{0.285714285714,0.992156863213,0.752941191196,0.525490224361});
-    out.push_back(std::vector<double>{0.428571428571,1.0,1.0,0.600000023842});
-    out.push_back(std::vector<double>{0.571428571429,0.219607844949,0.423529416323,0.690196096897});
-    out.push_back(std::vector<double>{0.714285714286,0.941176474094,0.00784313771874,0.498039215803});
-    out.push_back(std::vector<double>{0.857142857143,0.749019622803,0.35686275363,0.0901960805058});
-    out.push_back(std::vector<double>{1.0,0.40000000596,0.40000000596,0.40000000596});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Dark2(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.105882354081,0.61960786581,0.466666668653});
-    out.push_back(std::vector<double>{0.142857142857,0.850980401039,0.372549027205,0.00784313771874});
-    out.push_back(std::vector<double>{0.285714285714,0.458823531866,0.439215689898,0.701960802078});
-    out.push_back(std::vector<double>{0.428571428571,0.905882358551,0.160784319043,0.541176497936});
-    out.push_back(std::vector<double>{0.571428571429,0.40000000596,0.65098041296,0.117647059262});
-    out.push_back(std::vector<double>{0.714285714286,0.901960790157,0.670588254929,0.00784313771874});
-    out.push_back(std::vector<double>{0.857142857143,0.65098041296,0.46274510026,0.113725490868});
-    out.push_back(std::vector<double>{1.0,0.40000000596,0.40000000596,0.40000000596});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Paired(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.65098041296,0.807843148708,0.890196084976});
-    out.push_back(std::vector<double>{0.0909090909091,0.121568627656,0.470588237047,0.705882370472});
-    out.push_back(std::vector<double>{0.181818181818,0.698039233685,0.874509811401,0.541176497936});
-    out.push_back(std::vector<double>{0.272727272727,0.20000000298,0.627451002598,0.172549024224});
-    out.push_back(std::vector<double>{0.363636363636,0.984313726425,0.603921592236,0.600000023842});
-    out.push_back(std::vector<double>{0.454545454545,0.890196084976,0.101960785687,0.109803922474});
-    out.push_back(std::vector<double>{0.545454545455,0.992156863213,0.749019622803,0.435294121504});
-    out.push_back(std::vector<double>{0.636363636364,1.0,0.498039215803,0.0});
-    out.push_back(std::vector<double>{0.727272727273,0.792156875134,0.698039233685,0.839215695858});
-    out.push_back(std::vector<double>{0.818181818182,0.415686279535,0.239215686917,0.603921592236});
-    out.push_back(std::vector<double>{0.909090909091,1.0,1.0,0.600000023842});
-    out.push_back(std::vector<double>{1.0,0.694117665291,0.349019616842,0.156862750649});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Pastel1(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.984313726425,0.705882370472,0.68235296011});
-    out.push_back(std::vector<double>{0.125,0.701960802078,0.803921580315,0.890196084976});
-    out.push_back(std::vector<double>{0.25,0.800000011921,0.921568632126,0.772549033165});
-    out.push_back(std::vector<double>{0.375,0.870588243008,0.796078443527,0.89411765337});
-    out.push_back(std::vector<double>{0.5,0.996078431606,0.850980401039,0.65098041296});
-    out.push_back(std::vector<double>{0.625,1.0,1.0,0.800000011921});
-    out.push_back(std::vector<double>{0.75,0.898039221764,0.847058832645,0.741176486015});
-    out.push_back(std::vector<double>{0.875,0.992156863213,0.854901969433,0.92549020052});
-    out.push_back(std::vector<double>{1.0,0.949019610882,0.949019610882,0.949019610882});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Pastel2(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.701960802078,0.886274516582,0.803921580315});
-    out.push_back(std::vector<double>{0.142857142857,0.992156863213,0.803921580315,0.674509823322});
-    out.push_back(std::vector<double>{0.285714285714,0.796078443527,0.835294127464,0.909803926945});
-    out.push_back(std::vector<double>{0.428571428571,0.956862747669,0.792156875134,0.89411765337});
-    out.push_back(std::vector<double>{0.571428571429,0.901960790157,0.960784316063,0.78823530674});
-    out.push_back(std::vector<double>{0.714285714286,1.0,0.949019610882,0.68235296011});
-    out.push_back(std::vector<double>{0.857142857143,0.945098042488,0.886274516582,0.800000011921});
-    out.push_back(std::vector<double>{1.0,0.800000011921,0.800000011921,0.800000011921});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Set1(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.89411765337,0.101960785687,0.109803922474});
-    out.push_back(std::vector<double>{0.125,0.215686276555,0.494117647409,0.721568644047});
-    out.push_back(std::vector<double>{0.25,0.301960796118,0.686274528503,0.290196090937});
-    out.push_back(std::vector<double>{0.375,0.596078455448,0.305882364511,0.639215707779});
-    out.push_back(std::vector<double>{0.5,1.0,0.498039215803,0.0});
-    out.push_back(std::vector<double>{0.625,1.0,1.0,0.20000000298});
-    out.push_back(std::vector<double>{0.75,0.65098041296,0.337254911661,0.156862750649});
-    out.push_back(std::vector<double>{0.875,0.96862745285,0.505882382393,0.749019622803});
-    out.push_back(std::vector<double>{1.0,0.600000023842,0.600000023842,0.600000023842});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Set2(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.40000000596,0.760784327984,0.647058844566});
-    out.push_back(std::vector<double>{0.142857142857,0.988235294819,0.552941203117,0.384313732386});
-    out.push_back(std::vector<double>{0.285714285714,0.552941203117,0.627451002598,0.796078443527});
-    out.push_back(std::vector<double>{0.428571428571,0.905882358551,0.541176497936,0.764705896378});
-    out.push_back(std::vector<double>{0.571428571429,0.65098041296,0.847058832645,0.329411774874});
-    out.push_back(std::vector<double>{0.714285714286,1.0,0.850980401039,0.184313729405});
-    out.push_back(std::vector<double>{0.857142857143,0.898039221764,0.768627464771,0.580392181873});
-    out.push_back(std::vector<double>{1.0,0.701960802078,0.701960802078,0.701960802078});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_Set3(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.552941203117,0.827450990677,0.780392169952});
-    out.push_back(std::vector<double>{0.0909090909091,1.0,1.0,0.701960802078});
-    out.push_back(std::vector<double>{0.181818181818,0.745098054409,0.729411780834,0.854901969433});
-    out.push_back(std::vector<double>{0.272727272727,0.984313726425,0.501960813999,0.447058826685});
-    out.push_back(std::vector<double>{0.363636363636,0.501960813999,0.694117665291,0.827450990677});
-    out.push_back(std::vector<double>{0.454545454545,0.992156863213,0.705882370472,0.384313732386});
-    out.push_back(std::vector<double>{0.545454545455,0.701960802078,0.870588243008,0.411764711142});
-    out.push_back(std::vector<double>{0.636363636364,0.988235294819,0.803921580315,0.898039221764});
-    out.push_back(std::vector<double>{0.727272727273,0.850980401039,0.850980401039,0.850980401039});
-    out.push_back(std::vector<double>{0.818181818182,0.737254917622,0.501960813999,0.741176486015});
-    out.push_back(std::vector<double>{0.909090909091,0.800000011921,0.921568632126,0.772549033165});
-    out.push_back(std::vector<double>{1.0,1.0,0.929411768913,0.435294121504});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_gist_earth(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,0.0});
-    out.push_back(std::vector<double>{0.0039,0.00259907932011,0.0,0.1684});
-    out.push_back(std::vector<double>{0.0078,0.00519815864023,0.0,0.2212});
-    out.push_back(std::vector<double>{0.0275,0.0183268413598,0.0,0.4329});
-    out.push_back(std::vector<double>{0.0314,0.0209259206799,0.00897047387606,0.4549});
-    out.push_back(std::vector<double>{0.1098,0.0731740793201,0.1893,0.469111952191});
-    out.push_back(std::vector<double>{0.1647,0.10976111898,0.3035,0.479063944223});
-    out.push_back(std::vector<double>{0.2078,0.13848427762,0.3841,0.48687689243});
-    out.push_back(std::vector<double>{0.2824,0.1882,0.502,0.5004});
-    out.push_back(std::vector<double>{0.4588,0.2714,0.603547993311,0.28447032013});
-    out.push_back(std::vector<double>{0.4667,0.288960421286,0.608095777592,0.2748});
-    out.push_back(std::vector<double>{0.5216,0.410994235033,0.6397,0.306801658163});
-    out.push_back(std::vector<double>{0.5451,0.463230931264,0.65001122449,0.3205});
-    out.push_back(std::vector<double>{0.549,0.4719,0.65172244898,0.321732608696});
-    out.push_back(std::vector<double>{0.698,0.7176,0.7171,0.36882458194});
-    out.push_back(std::vector<double>{0.7843,0.753669955654,0.642568181818,0.3961});
-    out.push_back(std::vector<double>{0.7882,0.7553,0.6392,0.405654644809});
-    out.push_back(std::vector<double>{0.7922,0.759774032106,0.6413,0.41545428051});
-    out.push_back(std::vector<double>{0.8,0.768498394712,0.6447,0.434563570128});
-    out.push_back(std::vector<double>{0.8078,0.777222757318,0.6481,0.453672859745});
-    out.push_back(std::vector<double>{0.8157,0.786058970727,0.6549,0.473027140255});
-    out.push_back(std::vector<double>{0.8667,0.843102880076,0.6991,0.597972495446});
-    out.push_back(std::vector<double>{0.8745,0.851827242682,0.7103,0.617081785064});
-    out.push_back(std::vector<double>{0.8824,0.860663456091,0.7216,0.636436065574});
-    out.push_back(std::vector<double>{0.8902,0.869387818697,0.7323,0.655545355191});
-    out.push_back(std::vector<double>{0.8941,0.87375,0.73765,0.6651});
-    out.push_back(std::vector<double>{0.898,0.878112181303,0.743,0.676855240793});
-    out.push_back(std::vector<double>{0.9412,0.926431728045,0.8275,0.80706713881});
-    out.push_back(std::vector<double>{0.9569,0.94399230406,0.8635,0.854389518414});
-    out.push_back(std::vector<double>{0.9647,0.952716666667,0.8816,0.8779});
-    out.push_back(std::vector<double>{0.9961,0.987837818697,0.9733,0.972544759207});
-    out.push_back(std::vector<double>{1.0,0.9922,0.9843,0.9843});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_terrain(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.2,0.2,0.6,1.0});
-    out.push_back(std::vector<double>{0.15,0.0,0.6,1.0,1.0});
-    out.push_back(std::vector<double>{0.25,0.0,0.8,0.4,1.0});
-    out.push_back(std::vector<double>{0.5,1.0,1.0,0.6,1.0});
-    out.push_back(std::vector<double>{0.75,0.5,0.36,0.33,1.0});
-    out.push_back(std::vector<double>{1.0,1.0,1.0,1.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_ocean(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,-2.0,0.5,0.0});
-    out.push_back(std::vector<double>{0.01,-1.97,0.485,0.01});
-    out.push_back(std::vector<double>{0.02,-1.94,0.47,0.02});
-    out.push_back(std::vector<double>{0.03,-1.91,0.455,0.03});
-    out.push_back(std::vector<double>{0.04,-1.88,0.44,0.04});
-    out.push_back(std::vector<double>{0.05,-1.85,0.425,0.05});
-    out.push_back(std::vector<double>{0.06,-1.82,0.41,0.06});
-    out.push_back(std::vector<double>{0.07,-1.79,0.395,0.07});
-    out.push_back(std::vector<double>{0.08,-1.76,0.38,0.08});
-    out.push_back(std::vector<double>{0.09,-1.73,0.365,0.09});
-    out.push_back(std::vector<double>{0.1,-1.7,0.35,0.1});
-    out.push_back(std::vector<double>{0.11,-1.67,0.335,0.11});
-    out.push_back(std::vector<double>{0.12,-1.64,0.32,0.12});
-    out.push_back(std::vector<double>{0.13,-1.61,0.305,0.13});
-    out.push_back(std::vector<double>{0.14,-1.58,0.29,0.14});
-    out.push_back(std::vector<double>{0.15,-1.55,0.275,0.15});
-    out.push_back(std::vector<double>{0.16,-1.52,0.26,0.16});
-    out.push_back(std::vector<double>{0.17,-1.49,0.245,0.17});
-    out.push_back(std::vector<double>{0.18,-1.46,0.23,0.18});
-    out.push_back(std::vector<double>{0.19,-1.43,0.215,0.19});
-    out.push_back(std::vector<double>{0.2,-1.4,0.2,0.2});
-    out.push_back(std::vector<double>{0.21,-1.37,0.185,0.21});
-    out.push_back(std::vector<double>{0.22,-1.34,0.17,0.22});
-    out.push_back(std::vector<double>{0.23,-1.31,0.155,0.23});
-    out.push_back(std::vector<double>{0.24,-1.28,0.14,0.24});
-    out.push_back(std::vector<double>{0.25,-1.25,0.125,0.25});
-    out.push_back(std::vector<double>{0.26,-1.22,0.11,0.26});
-    out.push_back(std::vector<double>{0.27,-1.19,0.095,0.27});
-    out.push_back(std::vector<double>{0.28,-1.16,0.08,0.28});
-    out.push_back(std::vector<double>{0.29,-1.13,0.065,0.29});
-    out.push_back(std::vector<double>{0.3,-1.1,0.05,0.3});
-    out.push_back(std::vector<double>{0.31,-1.07,0.035,0.31});
-    out.push_back(std::vector<double>{0.32,-1.04,0.02,0.32});
-    out.push_back(std::vector<double>{0.33,-1.01,0.005,0.33});
-    out.push_back(std::vector<double>{0.34,-0.98,0.01,0.34});
-    out.push_back(std::vector<double>{0.35,-0.95,0.025,0.35});
-    out.push_back(std::vector<double>{0.36,-0.92,0.04,0.36});
-    out.push_back(std::vector<double>{0.37,-0.89,0.055,0.37});
-    out.push_back(std::vector<double>{0.38,-0.86,0.07,0.38});
-    out.push_back(std::vector<double>{0.39,-0.83,0.085,0.39});
-    out.push_back(std::vector<double>{0.4,-0.8,0.1,0.4});
-    out.push_back(std::vector<double>{0.41,-0.77,0.115,0.41});
-    out.push_back(std::vector<double>{0.42,-0.74,0.13,0.42});
-    out.push_back(std::vector<double>{0.43,-0.71,0.145,0.43});
-    out.push_back(std::vector<double>{0.44,-0.68,0.16,0.44});
-    out.push_back(std::vector<double>{0.45,-0.65,0.175,0.45});
-    out.push_back(std::vector<double>{0.46,-0.62,0.19,0.46});
-    out.push_back(std::vector<double>{0.47,-0.59,0.205,0.47});
-    out.push_back(std::vector<double>{0.48,-0.56,0.22,0.48});
-    out.push_back(std::vector<double>{0.49,-0.53,0.235,0.49});
-    out.push_back(std::vector<double>{0.5,-0.5,0.25,0.5});
-    out.push_back(std::vector<double>{0.51,-0.47,0.265,0.51});
-    out.push_back(std::vector<double>{0.52,-0.44,0.28,0.52});
-    out.push_back(std::vector<double>{0.53,-0.41,0.295,0.53});
-    out.push_back(std::vector<double>{0.54,-0.38,0.31,0.54});
-    out.push_back(std::vector<double>{0.55,-0.35,0.325,0.55});
-    out.push_back(std::vector<double>{0.56,-0.32,0.34,0.56});
-    out.push_back(std::vector<double>{0.57,-0.29,0.355,0.57});
-    out.push_back(std::vector<double>{0.58,-0.26,0.37,0.58});
-    out.push_back(std::vector<double>{0.59,-0.23,0.385,0.59});
-    out.push_back(std::vector<double>{0.6,-0.2,0.4,0.6});
-    out.push_back(std::vector<double>{0.61,-0.17,0.415,0.61});
-    out.push_back(std::vector<double>{0.62,-0.14,0.43,0.62});
-    out.push_back(std::vector<double>{0.63,-0.11,0.445,0.63});
-    out.push_back(std::vector<double>{0.64,-0.08,0.46,0.64});
-    out.push_back(std::vector<double>{0.65,-0.05,0.475,0.65});
-    out.push_back(std::vector<double>{0.66,-0.02,0.49,0.66});
-    out.push_back(std::vector<double>{0.67,0.01,0.505,0.67});
-    out.push_back(std::vector<double>{0.68,0.04,0.52,0.68});
-    out.push_back(std::vector<double>{0.69,0.07,0.535,0.69});
-    out.push_back(std::vector<double>{0.7,0.1,0.55,0.7});
-    out.push_back(std::vector<double>{0.71,0.13,0.565,0.71});
-    out.push_back(std::vector<double>{0.72,0.16,0.58,0.72});
-    out.push_back(std::vector<double>{0.73,0.19,0.595,0.73});
-    out.push_back(std::vector<double>{0.74,0.22,0.61,0.74});
-    out.push_back(std::vector<double>{0.75,0.25,0.625,0.75});
-    out.push_back(std::vector<double>{0.76,0.28,0.64,0.76});
-    out.push_back(std::vector<double>{0.77,0.31,0.655,0.77});
-    out.push_back(std::vector<double>{0.78,0.34,0.67,0.78});
-    out.push_back(std::vector<double>{0.79,0.37,0.685,0.79});
-    out.push_back(std::vector<double>{0.8,0.4,0.7,0.8});
-    out.push_back(std::vector<double>{0.81,0.43,0.715,0.81});
-    out.push_back(std::vector<double>{0.82,0.46,0.73,0.82});
-    out.push_back(std::vector<double>{0.83,0.49,0.745,0.83});
-    out.push_back(std::vector<double>{0.84,0.52,0.76,0.84});
-    out.push_back(std::vector<double>{0.85,0.55,0.775,0.85});
-    out.push_back(std::vector<double>{0.86,0.58,0.79,0.86});
-    out.push_back(std::vector<double>{0.87,0.61,0.805,0.87});
-    out.push_back(std::vector<double>{0.88,0.64,0.82,0.88});
-    out.push_back(std::vector<double>{0.89,0.67,0.835,0.89});
-    out.push_back(std::vector<double>{0.9,0.7,0.85,0.9});
-    out.push_back(std::vector<double>{0.91,0.73,0.865,0.91});
-    out.push_back(std::vector<double>{0.92,0.76,0.88,0.92});
-    out.push_back(std::vector<double>{0.93,0.79,0.895,0.93});
-    out.push_back(std::vector<double>{0.94,0.82,0.91,0.94});
-    out.push_back(std::vector<double>{0.95,0.85,0.925,0.95});
-    out.push_back(std::vector<double>{0.96,0.88,0.94,0.96});
-    out.push_back(std::vector<double>{0.97,0.91,0.955,0.97});
-    out.push_back(std::vector<double>{0.98,0.94,0.97,0.98});
-    out.push_back(std::vector<double>{0.99,0.97,0.985,0.99});
-    out.push_back(std::vector<double>{1.0,1.0,1.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_gist_stern(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0,0.0});
-    out.push_back(std::vector<double>{0.0547,1.0,0.0547,0.1094});
-    out.push_back(std::vector<double>{0.25,0.027,0.25,0.5});
-    out.push_back(std::vector<double>{0.5,0.351333333333,0.5,1.0});
-    out.push_back(std::vector<double>{0.735,0.656206666667,0.735,0.0});
-    out.push_back(std::vector<double>{1.0,1.0,1,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_brg(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,1.0,1.0});
-    out.push_back(std::vector<double>{0.5,1.0,0.0,0.0,1.0});
-    out.push_back(std::vector<double>{1.0,0.0,1.0,0.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_CMRmap(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,0.0});
-    out.push_back(std::vector<double>{0.125,0.15,0.15,0.5});
-    out.push_back(std::vector<double>{0.25,0.3,0.15,0.75});
-    out.push_back(std::vector<double>{0.375,0.6,0.2,0.5});
-    out.push_back(std::vector<double>{0.5,1.0,0.25,0.15});
-    out.push_back(std::vector<double>{0.625,0.9,0.5,0.0});
-    out.push_back(std::vector<double>{0.75,0.9,0.75,0.1});
-    out.push_back(std::vector<double>{0.875,0.9,0.9,0.5});
-    out.push_back(std::vector<double>{1.0,1.0,1.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_cubehelix(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,0.0});
-    out.push_back(std::vector<double>{0.01,0.0167671371066,0.00550435206624,0.015657290631});
-    out.push_back(std::vector<double>{0.02,0.0322739437462,0.0114035998583,0.0326337645351});
-    out.push_back(std::vector<double>{0.03,0.0463930673944,0.0178017204642,0.0507190490354});
-    out.push_back(std::vector<double>{0.04,0.0590252597162,0.0247914047082,0.0696866592582});
-    out.push_back(std::vector<double>{0.05,0.0701003683596,0.0324528329636,0.0892978593702});
-    out.push_back(std::vector<double>{0.06,0.0795778939531,0.040852652808,0.109305627007});
-    out.push_back(std::vector<double>{0.07,0.0874471122166,0.0500431688988,0.129458665482});
-    out.push_back(std::vector<double>{0.08,0.0937267667529,0.0600617525239,0.149505408604});
-    out.push_back(std::vector<double>{0.09,0.0984643436015,0.0709304753068,0.169197963851});
-    out.push_back(std::vector<double>{0.1,0.101734943932,0.0826559685538,0.188295941271});
-    out.push_back(std::vector<double>{0.11,0.10363977627,0.0952295067576,0.206570117709});
-    out.push_back(std::vector<double>{0.12,0.104304294344,0.10862731086,0.223805888827});
-    out.push_back(std::vector<double>{0.13,0.10387601091,0.122811064054,0.239806464822});
-    out.push_back(std::vector<double>{0.14,0.102522021785,0.137728630208,0.254395769663});
-    out.push_back(std::vector<double>{0.15,0.100426277692,0.153314962472,0.267421008124});
-    out.push_back(std::vector<double>{0.16,0.0977866443367,0.16949318725,0.278754869649});
-    out.push_back(std::vector<double>{0.17,0.0948117934899,0.18617584663,0.288297343288});
-    out.push_back(std::vector<double>{0.18,0.0917179695361,0.203266280434,0.295977123351});
-    out.push_back(std::vector<double>{0.19,0.0887256771128,0.220660127452,0.301752591034});
-    out.push_back(std::vector<double>{0.2,0.0860563360058,0.238246924042,0.305612363081});
-    out.push_back(std::vector<double>{0.21,0.0839289494145,0.255911777249,0.307575404304});
-    out.push_back(std::vector<double>{0.22,0.0825568310556,0.273537088819,0.307690706635});
-    out.push_back(std::vector<double>{0.23,0.0821444353558,0.291004306065,0.306036543042});
-    out.push_back(std::vector<double>{0.24,0.0828843332025,0.308195675373,0.302719310236});
-    out.push_back(std::vector<double>{0.25,0.0849543734101,0.324995974371,0.297871979392});
-    out.push_back(std::vector<double>{0.26,0.0885150672564,0.341294199229,0.291652179136});
-    out.push_back(std::vector<double>{0.27,0.0937072301719,0.356985184363,0.28423993974});
-    out.push_back(std::vector<double>{0.28,0.100649910989,0.371971132909,0.275835131701});
-    out.push_back(std::vector<double>{0.29,0.109438635112,0.386163037646,0.266654635731});
-    out.push_back(std::vector<double>{0.3,0.120143983612,0.399481973689,0.256929284415});
-    out.push_back(std::vector<double>{0.31,0.132810525647,0.411860246051,0.246900618625});
-    out.push_back(std::vector<double>{0.32,0.147456116795,0.42324237727,0.236817503877});
-    out.push_back(std::vector<double>{0.33,0.16407157096,0.433585922462,0.226932653467});
-    out.push_back(std::vector<double>{0.34,0.182620708501,0.442862101547,0.217499106141});
-    out.push_back(std::vector<double>{0.35,0.203040778228,0.451056240887,0.208766706399});
-    out.push_back(std::vector<double>{0.36,0.225243245948,0.458168019136,0.200978635273});
-    out.push_back(std::vector<double>{0.37,0.24911493742,0.46421151472,0.194368038484});
-    out.push_back(std::vector<double>{0.38,0.274519518924,0.469215055032,0.189154797385});
-    out.push_back(std::vector<double>{0.39,0.301299294227,0.473220870048,0.185542486024});
-    out.push_back(std::vector<double>{0.4,0.329277292634,0.476284555658,0.183715554976});
-    out.push_back(std::vector<double>{0.41,0.358259619006,0.47847435451,0.183836779482});
-    out.push_back(std::vector<double>{0.42,0.38803803327,0.479870264578,0.186045005755});
-    out.push_back(std::vector<double>{0.43,0.418392723993,0.480562987906,0.190453225261});
-    out.push_back(std::vector<double>{0.44,0.449095238101,0.480652734079,0.197147002362});
-    out.push_back(std::vector<double>{0.45,0.479911526849,0.480247894852,0.206183275906});
-    out.push_back(std::vector<double>{0.46,0.510605066717,0.479463608068,0.217589550391});
-    out.push_back(std::vector<double>{0.47,0.540940012937,0.478420230404,0.231363487074});
-    out.push_back(std::vector<double>{0.48,0.570684343051,0.47724173968,0.247472900096});
-    out.push_back(std::vector<double>{0.49,0.599612948049,0.476054088374,0.265856157292});
-    out.push_back(std::vector<double>{0.5,0.627510629407,0.474983530596,0.286422979982});
-    out.push_back(std::vector<double>{0.51,0.654174961606,0.474154945127,0.309055630749});
-    out.push_back(std::vector<double>{0.52,0.679418981537,0.47369017716,0.333610473028});
-    out.push_back(std::vector<double>{0.53,0.703073668519,0.473706421117,0.359919881447});
-    out.push_back(std::vector<double>{0.54,0.724990181403,0.474314666388,0.387794477139});
-    out.push_back(std::vector<double>{0.55,0.745041822507,0.475618226994,0.417025657954});
-    out.push_back(std::vector<double>{0.56,0.763125701699,0.477711375077,0.447388389562});
-    out.push_back(std::vector<double>{0.57,0.779164077933,0.480678096747,0.478644219902});
-    out.push_back(std::vector<double>{0.58,0.793105359808,0.484590987227,0.510544476476});
-    out.push_back(std::vector<double>{0.59,0.804924751224,0.489510300384,0.542833603455});
-    out.push_back(std::vector<double>{0.6,0.81462453292,0.495483165723,0.575252593642});
-    out.push_back(std::vector<double>{0.61,0.822233975487,0.502542983706,0.607542469001});
-    out.push_back(std::vector<double>{0.62,0.827808884384,0.510709007907,0.639447762706});
-    out.push_back(std::vector<double>{0.63,0.831430782341,0.51998612007,0.6707199555});
-    out.push_back(std::vector<double>{0.64,0.833205739418,0.530364801532,0.701120819686});
-    out.push_back(std::vector<double>{0.65,0.833262865691,0.541821301937,0.730425625095});
-    out.push_back(std::vector<double>{0.66,0.831752486096,0.55431800348,0.758426163113});
-    out.push_back(std::vector<double>{0.67,0.828844021268,0.567803976334,0.78493354708});
-    out.push_back(std::vector<double>{0.68,0.824723602224,0.58221571835,0.809780750211});
-    out.push_back(std::vector<double>{0.69,0.819591450415,0.597478069626,0.832824845508});
-    out.push_back(std::vector<double>{0.7,0.813659057951,0.613505290142,0.853948915958});
-    out.push_back(std::vector<double>{0.71,0.807146205626,0.630202286478,0.873063607543});
-    out.push_back(std::vector<double>{0.72,0.800277858743,0.647465971503,0.890108302235});
-    out.push_back(std::vector<double>{0.73,0.793280982608,0.665186739111,0.905051893056});
-    out.push_back(std::vector<double>{0.74,0.786381320848,0.683250034408,0.917893148547});
-    out.push_back(std::vector<double>{0.75,0.779800180499,0.701537998365,0.928660659349});
-    out.push_back(std::vector<double>{0.76,0.773751268013,0.719931164818,0.93741236515});
-    out.push_back(std::vector<double>{0.77,0.768437619924,0.738310186852,0.944234665867});
-    out.push_back(std::vector<double>{0.78,0.764048671012,0.756557569032,0.949241126482});
-    out.push_back(std::vector<double>{0.79,0.760757501252,0.77455938169,0.952570790482});
-    out.push_back(std::vector<double>{0.8,0.758718300801,0.792206933547,0.954386122191});
-    out.push_back(std::vector<double>{0.81,0.758064089681,0.809398379272,0.954870603413});
-    out.push_back(std::vector<double>{0.82,0.758904725718,0.826040239276,0.954226014649});
-    out.push_back(std::vector<double>{0.83,0.761325230775,0.842048809994,0.952669435648});
-    out.push_back(std::vector<double>{0.84,0.76538446133,0.857351444171,0.950430004114});
-    out.push_back(std::vector<double>{0.85,0.771114145101,0.871887682184,0.947745475016});
-    out.push_back(std::vector<double>{0.86,0.778518300771,0.885610217266,0.944858626059});
-    out.push_back(std::vector<double>{0.87,0.787573052913,0.898485679491,0.942013557373});
-    out.push_back(std::vector<double>{0.88,0.798226849084,0.910495225666,0.939451935465});
-    out.push_back(std::vector<double>{0.89,0.810401080762,0.921634924703,0.937409232749});
-    out.push_back(std::vector<double>{0.9,0.823991104438,0.931915930645,0.936111014667});
-    out.push_back(std::vector<double>{0.91,0.8388676538,0.941364438278,0.935769326388});
-    out.push_back(std::vector<double>{0.92,0.854878628608,0.950021419056,0.936579230417});
-    out.push_back(std::vector<double>{0.93,0.871851240679,0.957942138006,0.938715545082});
-    out.push_back(std::vector<double>{0.94,0.889594492366,0.965195455166,0.942329831894});
-    out.push_back(std::vector<double>{0.95,0.907901958181,0.971862918026,0.947547677089});
-    out.push_back(std::vector<double>{0.96,0.926554835747,0.978037654346,0.954466309433});
-    out.push_back(std::vector<double>{0.97,0.945325228225,0.983823077446,0.963152592493});
-    out.push_back(std::vector<double>{0.98,0.963979616706,0.989331418789,0.973641425212});
-    out.push_back(std::vector<double>{0.99,0.982282477921,0.994682105147,0.98593457972});
-    out.push_back(std::vector<double>{1.0,1.0,1.0,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_gnuplot(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,0.0});
-    out.push_back(std::vector<double>{0.01,0.1,1e-06,0.0627905195293});
-    out.push_back(std::vector<double>{0.02,0.141421356237,8e-06,0.125333233564});
-    out.push_back(std::vector<double>{0.03,0.173205080757,2.7e-05,0.187381314586});
-    out.push_back(std::vector<double>{0.04,0.2,6.4e-05,0.248689887165});
-    out.push_back(std::vector<double>{0.05,0.22360679775,0.000125,0.309016994375});
-    out.push_back(std::vector<double>{0.06,0.244948974278,0.000216,0.368124552685});
-    out.push_back(std::vector<double>{0.07,0.264575131106,0.000343,0.425779291565});
-    out.push_back(std::vector<double>{0.08,0.282842712475,0.000512,0.481753674102});
-    out.push_back(std::vector<double>{0.09,0.3,0.000729,0.535826794979});
-    out.push_back(std::vector<double>{0.1,0.316227766017,0.001,0.587785252292});
-    out.push_back(std::vector<double>{0.11,0.331662479036,0.001331,0.637423989749});
-    out.push_back(std::vector<double>{0.12,0.346410161514,0.001728,0.684547105929});
-    out.push_back(std::vector<double>{0.13,0.360555127546,0.002197,0.728968627421});
-    out.push_back(std::vector<double>{0.14,0.374165738677,0.002744,0.770513242776});
-    out.push_back(std::vector<double>{0.15,0.387298334621,0.003375,0.809016994375});
-    out.push_back(std::vector<double>{0.16,0.4,0.004096,0.844327925502});
-    out.push_back(std::vector<double>{0.17,0.412310562562,0.004913,0.876306680044});
-    out.push_back(std::vector<double>{0.18,0.424264068712,0.005832,0.904827052466});
-    out.push_back(std::vector<double>{0.19,0.435889894354,0.006859,0.929776485888});
-    out.push_back(std::vector<double>{0.2,0.4472135955,0.008,0.951056516295});
-    out.push_back(std::vector<double>{0.21,0.458257569496,0.009261,0.968583161129});
-    out.push_back(std::vector<double>{0.22,0.469041575982,0.010648,0.982287250729});
-    out.push_back(std::vector<double>{0.23,0.479583152331,0.012167,0.992114701314});
-    out.push_back(std::vector<double>{0.24,0.489897948557,0.013824,0.998026728428});
-    out.push_back(std::vector<double>{0.25,0.5,0.015625,1.0});
-    out.push_back(std::vector<double>{0.26,0.509901951359,0.017576,0.998026728428});
-    out.push_back(std::vector<double>{0.27,0.519615242271,0.019683,0.992114701314});
-    out.push_back(std::vector<double>{0.28,0.529150262213,0.021952,0.982287250729});
-    out.push_back(std::vector<double>{0.29,0.538516480713,0.024389,0.968583161129});
-    out.push_back(std::vector<double>{0.3,0.547722557505,0.027,0.951056516295});
-    out.push_back(std::vector<double>{0.31,0.556776436283,0.029791,0.929776485888});
-    out.push_back(std::vector<double>{0.32,0.565685424949,0.032768,0.904827052466});
-    out.push_back(std::vector<double>{0.33,0.574456264654,0.035937,0.876306680044});
-    out.push_back(std::vector<double>{0.34,0.583095189485,0.039304,0.844327925502});
-    out.push_back(std::vector<double>{0.35,0.59160797831,0.042875,0.809016994375});
-    out.push_back(std::vector<double>{0.36,0.6,0.046656,0.770513242776});
-    out.push_back(std::vector<double>{0.37,0.60827625303,0.050653,0.728968627421});
-    out.push_back(std::vector<double>{0.38,0.616441400297,0.054872,0.684547105929});
-    out.push_back(std::vector<double>{0.39,0.62449979984,0.059319,0.637423989749});
-    out.push_back(std::vector<double>{0.4,0.632455532034,0.064,0.587785252292});
-    out.push_back(std::vector<double>{0.41,0.640312423743,0.068921,0.535826794979});
-    out.push_back(std::vector<double>{0.42,0.648074069841,0.074088,0.481753674102});
-    out.push_back(std::vector<double>{0.43,0.65574385243,0.079507,0.425779291565});
-    out.push_back(std::vector<double>{0.44,0.663324958071,0.085184,0.368124552685});
-    out.push_back(std::vector<double>{0.45,0.67082039325,0.091125,0.309016994375});
-    out.push_back(std::vector<double>{0.46,0.678232998313,0.097336,0.248689887165});
-    out.push_back(std::vector<double>{0.47,0.68556546004,0.103823,0.187381314586});
-    out.push_back(std::vector<double>{0.48,0.692820323028,0.110592,0.125333233564});
-    out.push_back(std::vector<double>{0.49,0.7,0.117649,0.0627905195293});
-    out.push_back(std::vector<double>{0.5,0.707106781187,0.125,1.22464679915e-16});
-    out.push_back(std::vector<double>{0.51,0.714142842854,0.132651,-0.0627905195293});
-    out.push_back(std::vector<double>{0.52,0.721110255093,0.140608,-0.125333233564});
-    out.push_back(std::vector<double>{0.53,0.728010988928,0.148877,-0.187381314586});
-    out.push_back(std::vector<double>{0.54,0.734846922835,0.157464,-0.248689887165});
-    out.push_back(std::vector<double>{0.55,0.74161984871,0.166375,-0.309016994375});
-    out.push_back(std::vector<double>{0.56,0.748331477355,0.175616,-0.368124552685});
-    out.push_back(std::vector<double>{0.57,0.754983443527,0.185193,-0.425779291565});
-    out.push_back(std::vector<double>{0.58,0.761577310586,0.195112,-0.481753674102});
-    out.push_back(std::vector<double>{0.59,0.768114574787,0.205379,-0.535826794979});
-    out.push_back(std::vector<double>{0.6,0.774596669241,0.216,-0.587785252292});
-    out.push_back(std::vector<double>{0.61,0.781024967591,0.226981,-0.637423989749});
-    out.push_back(std::vector<double>{0.62,0.787400787401,0.238328,-0.684547105929});
-    out.push_back(std::vector<double>{0.63,0.793725393319,0.250047,-0.728968627421});
-    out.push_back(std::vector<double>{0.64,0.8,0.262144,-0.770513242776});
-    out.push_back(std::vector<double>{0.65,0.80622577483,0.274625,-0.809016994375});
-    out.push_back(std::vector<double>{0.66,0.812403840464,0.287496,-0.844327925502});
-    out.push_back(std::vector<double>{0.67,0.818535277187,0.300763,-0.876306680044});
-    out.push_back(std::vector<double>{0.68,0.824621125124,0.314432,-0.904827052466});
-    out.push_back(std::vector<double>{0.69,0.830662386292,0.328509,-0.929776485888});
-    out.push_back(std::vector<double>{0.7,0.836660026534,0.343,-0.951056516295});
-    out.push_back(std::vector<double>{0.71,0.842614977318,0.357911,-0.968583161129});
-    out.push_back(std::vector<double>{0.72,0.848528137424,0.373248,-0.982287250729});
-    out.push_back(std::vector<double>{0.73,0.854400374532,0.389017,-0.992114701314});
-    out.push_back(std::vector<double>{0.74,0.860232526704,0.405224,-0.998026728428});
-    out.push_back(std::vector<double>{0.75,0.866025403784,0.421875,-1.0});
-    out.push_back(std::vector<double>{0.76,0.871779788708,0.438976,-0.998026728428});
-    out.push_back(std::vector<double>{0.77,0.877496438739,0.456533,-0.992114701314});
-    out.push_back(std::vector<double>{0.78,0.883176086633,0.474552,-0.982287250729});
-    out.push_back(std::vector<double>{0.79,0.888819441732,0.493039,-0.968583161129});
-    out.push_back(std::vector<double>{0.8,0.894427191,0.512,-0.951056516295});
-    out.push_back(std::vector<double>{0.81,0.9,0.531441,-0.929776485888});
-    out.push_back(std::vector<double>{0.82,0.905538513814,0.551368,-0.904827052466});
-    out.push_back(std::vector<double>{0.83,0.911043357914,0.571787,-0.876306680044});
-    out.push_back(std::vector<double>{0.84,0.916515138991,0.592704,-0.844327925502});
-    out.push_back(std::vector<double>{0.85,0.921954445729,0.614125,-0.809016994375});
-    out.push_back(std::vector<double>{0.86,0.92736184955,0.636056,-0.770513242776});
-    out.push_back(std::vector<double>{0.87,0.932737905309,0.658503,-0.728968627421});
-    out.push_back(std::vector<double>{0.88,0.938083151965,0.681472,-0.684547105929});
-    out.push_back(std::vector<double>{0.89,0.943398113206,0.704969,-0.637423989749});
-    out.push_back(std::vector<double>{0.9,0.948683298051,0.729,-0.587785252292});
-    out.push_back(std::vector<double>{0.91,0.953939201417,0.753571,-0.535826794979});
-    out.push_back(std::vector<double>{0.92,0.959166304663,0.778688,-0.481753674102});
-    out.push_back(std::vector<double>{0.93,0.964365076099,0.804357,-0.425779291565});
-    out.push_back(std::vector<double>{0.94,0.969535971483,0.830584,-0.368124552685});
-    out.push_back(std::vector<double>{0.95,0.974679434481,0.857375,-0.309016994375});
-    out.push_back(std::vector<double>{0.96,0.979795897113,0.884736,-0.248689887165});
-    out.push_back(std::vector<double>{0.97,0.98488578018,0.912673,-0.187381314586});
-    out.push_back(std::vector<double>{0.98,0.989949493661,0.941192,-0.125333233564});
-    out.push_back(std::vector<double>{0.99,0.994987437107,0.970299,-0.0627905195293});
-    out.push_back(std::vector<double>{1.0,1.0,1.0,-2.44929359829e-16});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_gnuplot2(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,-0.78125,-0.84,0.0});
-    out.push_back(std::vector<double>{0.01,-0.75,-0.82,0.04});
-    out.push_back(std::vector<double>{0.02,-0.71875,-0.8,0.08});
-    out.push_back(std::vector<double>{0.03,-0.6875,-0.78,0.12});
-    out.push_back(std::vector<double>{0.04,-0.65625,-0.76,0.16});
-    out.push_back(std::vector<double>{0.05,-0.625,-0.74,0.2});
-    out.push_back(std::vector<double>{0.06,-0.59375,-0.72,0.24});
-    out.push_back(std::vector<double>{0.07,-0.5625,-0.7,0.28});
-    out.push_back(std::vector<double>{0.08,-0.53125,-0.68,0.32});
-    out.push_back(std::vector<double>{0.09,-0.5,-0.66,0.36});
-    out.push_back(std::vector<double>{0.1,-0.46875,-0.64,0.4});
-    out.push_back(std::vector<double>{0.11,-0.4375,-0.62,0.44});
-    out.push_back(std::vector<double>{0.12,-0.40625,-0.6,0.48});
-    out.push_back(std::vector<double>{0.13,-0.375,-0.58,0.52});
-    out.push_back(std::vector<double>{0.14,-0.34375,-0.56,0.56});
-    out.push_back(std::vector<double>{0.15,-0.3125,-0.54,0.6});
-    out.push_back(std::vector<double>{0.16,-0.28125,-0.52,0.64});
-    out.push_back(std::vector<double>{0.17,-0.25,-0.5,0.68});
-    out.push_back(std::vector<double>{0.18,-0.21875,-0.48,0.72});
-    out.push_back(std::vector<double>{0.19,-0.1875,-0.46,0.76});
-    out.push_back(std::vector<double>{0.2,-0.15625,-0.44,0.8});
-    out.push_back(std::vector<double>{0.21,-0.125,-0.42,0.84});
-    out.push_back(std::vector<double>{0.22,-0.09375,-0.4,0.88});
-    out.push_back(std::vector<double>{0.23,-0.0625,-0.38,0.92});
-    out.push_back(std::vector<double>{0.24,-0.03125,-0.36,0.96});
-    out.push_back(std::vector<double>{0.25,0.0,-0.34,1.34});
-    out.push_back(std::vector<double>{0.26,0.03125,-0.32,1.32});
-    out.push_back(std::vector<double>{0.27,0.0625,-0.3,1.3});
-    out.push_back(std::vector<double>{0.28,0.09375,-0.28,1.28});
-    out.push_back(std::vector<double>{0.29,0.125,-0.26,1.26});
-    out.push_back(std::vector<double>{0.3,0.15625,-0.24,1.24});
-    out.push_back(std::vector<double>{0.31,0.1875,-0.22,1.22});
-    out.push_back(std::vector<double>{0.32,0.21875,-0.2,1.2});
-    out.push_back(std::vector<double>{0.33,0.25,-0.18,1.18});
-    out.push_back(std::vector<double>{0.34,0.28125,-0.16,1.16});
-    out.push_back(std::vector<double>{0.35,0.3125,-0.14,1.14});
-    out.push_back(std::vector<double>{0.36,0.34375,-0.12,1.12});
-    out.push_back(std::vector<double>{0.37,0.375,-0.1,1.1});
-    out.push_back(std::vector<double>{0.38,0.40625,-0.08,1.08});
-    out.push_back(std::vector<double>{0.39,0.4375,-0.06,1.06});
-    out.push_back(std::vector<double>{0.4,0.46875,-0.04,1.04});
-    out.push_back(std::vector<double>{0.41,0.5,-0.02,1.02});
-    out.push_back(std::vector<double>{0.42,0.53125,0.0,1.0});
-    out.push_back(std::vector<double>{0.43,0.5625,0.02,0.98});
-    out.push_back(std::vector<double>{0.44,0.59375,0.04,0.96});
-    out.push_back(std::vector<double>{0.45,0.625,0.06,0.94});
-    out.push_back(std::vector<double>{0.46,0.65625,0.08,0.92});
-    out.push_back(std::vector<double>{0.47,0.6875,0.1,0.9});
-    out.push_back(std::vector<double>{0.48,0.71875,0.12,0.88});
-    out.push_back(std::vector<double>{0.49,0.75,0.14,0.86});
-    out.push_back(std::vector<double>{0.5,0.78125,0.16,0.84});
-    out.push_back(std::vector<double>{0.51,0.8125,0.18,0.82});
-    out.push_back(std::vector<double>{0.52,0.84375,0.2,0.8});
-    out.push_back(std::vector<double>{0.53,0.875,0.22,0.78});
-    out.push_back(std::vector<double>{0.54,0.90625,0.24,0.76});
-    out.push_back(std::vector<double>{0.55,0.9375,0.26,0.74});
-    out.push_back(std::vector<double>{0.56,0.96875,0.28,0.72});
-    out.push_back(std::vector<double>{0.57,1.0,0.3,0.7});
-    out.push_back(std::vector<double>{0.58,1.03125,0.32,0.68});
-    out.push_back(std::vector<double>{0.59,1.0625,0.34,0.66});
-    out.push_back(std::vector<double>{0.6,1.09375,0.36,0.64});
-    out.push_back(std::vector<double>{0.61,1.125,0.38,0.62});
-    out.push_back(std::vector<double>{0.62,1.15625,0.4,0.6});
-    out.push_back(std::vector<double>{0.63,1.1875,0.42,0.58});
-    out.push_back(std::vector<double>{0.64,1.21875,0.44,0.56});
-    out.push_back(std::vector<double>{0.65,1.25,0.46,0.54});
-    out.push_back(std::vector<double>{0.66,1.28125,0.48,0.52});
-    out.push_back(std::vector<double>{0.67,1.3125,0.5,0.5});
-    out.push_back(std::vector<double>{0.68,1.34375,0.52,0.48});
-    out.push_back(std::vector<double>{0.69,1.375,0.54,0.46});
-    out.push_back(std::vector<double>{0.7,1.40625,0.56,0.44});
-    out.push_back(std::vector<double>{0.71,1.4375,0.58,0.42});
-    out.push_back(std::vector<double>{0.72,1.46875,0.6,0.4});
-    out.push_back(std::vector<double>{0.73,1.5,0.62,0.38});
-    out.push_back(std::vector<double>{0.74,1.53125,0.64,0.36});
-    out.push_back(std::vector<double>{0.75,1.5625,0.66,0.34});
-    out.push_back(std::vector<double>{0.76,1.59375,0.68,0.32});
-    out.push_back(std::vector<double>{0.77,1.625,0.7,0.3});
-    out.push_back(std::vector<double>{0.78,1.65625,0.72,0.28});
-    out.push_back(std::vector<double>{0.79,1.6875,0.74,0.26});
-    out.push_back(std::vector<double>{0.8,1.71875,0.76,0.24});
-    out.push_back(std::vector<double>{0.81,1.75,0.78,0.22});
-    out.push_back(std::vector<double>{0.82,1.78125,0.8,0.2});
-    out.push_back(std::vector<double>{0.83,1.8125,0.82,0.18});
-    out.push_back(std::vector<double>{0.84,1.84375,0.84,0.16});
-    out.push_back(std::vector<double>{0.85,1.875,0.86,0.14});
-    out.push_back(std::vector<double>{0.86,1.90625,0.88,0.12});
-    out.push_back(std::vector<double>{0.87,1.9375,0.9,0.1});
-    out.push_back(std::vector<double>{0.88,1.96875,0.92,0.08});
-    out.push_back(std::vector<double>{0.89,2.0,0.94,0.06});
-    out.push_back(std::vector<double>{0.9,2.03125,0.96,0.04});
-    out.push_back(std::vector<double>{0.91,2.0625,0.98,0.02});
-    out.push_back(std::vector<double>{0.92,2.09375,1.0,0.0});
-    out.push_back(std::vector<double>{0.93,2.125,1.02,0.125});
-    out.push_back(std::vector<double>{0.94,2.15625,1.04,0.25});
-    out.push_back(std::vector<double>{0.95,2.1875,1.06,0.375});
-    out.push_back(std::vector<double>{0.96,2.21875,1.08,0.5});
-    out.push_back(std::vector<double>{0.97,2.25,1.1,0.625});
-    out.push_back(std::vector<double>{0.98,2.28125,1.12,0.75});
-    out.push_back(std::vector<double>{0.99,2.3125,1.14,0.875});
-    out.push_back(std::vector<double>{1.0,2.34375,1.16,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_gist_ncar(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,0.502});
-    out.push_back(std::vector<double>{0.051,0.0,0.3722,0.0222});
-    out.push_back(std::vector<double>{0.1059,0.0,0.0,0.935145918367});
-    out.push_back(std::vector<double>{0.1098,0.0,0.0550741176471,1.0});
-    out.push_back(std::vector<double>{0.1569,0.0,0.7202,1.0});
-    out.push_back(std::vector<double>{0.1608,0.0,0.7537,1.0});
-    out.push_back(std::vector<double>{0.1647,0.0,0.7752,1.0});
-    out.push_back(std::vector<double>{0.2039,0.0,0.94798745098,1.0});
-    out.push_back(std::vector<double>{0.2157,0.0,1.0,0.922637755102});
-    out.push_back(std::vector<double>{0.2588,0.0,0.9804,0.640068877551});
-    out.push_back(std::vector<double>{0.2627,0.0,0.9804,0.6145});
-    out.push_back(std::vector<double>{0.2706,0.0,0.9804,0.532079796265});
-    out.push_back(std::vector<double>{0.3098,0.0,0.996747234043,0.123108658744});
-    out.push_back(std::vector<double>{0.3176,0.0496736842105,1.0,0.0417317487267});
-    out.push_back(std::vector<double>{0.3216,0.0751473684211,0.984949019608,0.0});
-    out.push_back(std::vector<double>{0.3686,0.374463157895,0.8081,0.0});
-    out.push_back(std::vector<double>{0.3725,0.3993,0.820806451613,0.0});
-    out.push_back(std::vector<double>{0.4157,0.484852941176,0.96155483871,0.0});
-    out.push_back(std::vector<double>{0.4235,0.5003,0.986967741935,0.0310673469388});
-    out.push_back(std::vector<double>{0.4275,0.518504007286,1.0,0.0469993197279});
-    out.push_back(std::vector<double>{0.4745,0.732401092896,1.0,0.2342});
-    out.push_back(std::vector<double>{0.5216,0.946753278689,1.0,0.0466010204082});
-    out.push_back(std::vector<double>{0.5333,1.0,0.971144262295,0.0});
-    out.push_back(std::vector<double>{0.5804,1.0,0.854981420765,0.0});
-    out.push_back(std::vector<double>{0.6314,1.0,0.7292,0.0549});
-    out.push_back(std::vector<double>{0.6863,1.0,0.2796,0.00364132653061});
-    out.push_back(std::vector<double>{0.6902,1.0,0.261055102041,0.0});
-    out.push_back(std::vector<double>{0.7373,1.0,0.0370897959184,0.0});
-    out.push_back(std::vector<double>{0.7451,1.0,0.0,0.138354098361});
-    out.push_back(std::vector<double>{0.7922,1.0,0.0,0.9738});
-    out.push_back(std::vector<double>{0.8,0.946266666667,0.0268632612967,1.0});
-    out.push_back(std::vector<double>{0.8431,0.649355555556,0.1753,1.0});
-    out.push_back(std::vector<double>{0.8471,0.6218,0.198957559199,0.995198542805});
-    out.push_back(std::vector<double>{0.898,0.9235,0.5,0.9341});
-    out.push_back(std::vector<double>{1.0,0.9961,0.9725,0.9961});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_nipy_spectral(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.0,0.0,0.0});
-    out.push_back(std::vector<double>{0.05,0.4667,0.0,0.5333});
-    out.push_back(std::vector<double>{0.1,0.5333,0.0,0.6});
-    out.push_back(std::vector<double>{0.15,0.0,0.0,0.6667});
-    out.push_back(std::vector<double>{0.2,0.0,0.0,0.8667});
-    out.push_back(std::vector<double>{0.25,0.0,0.4667,0.8667});
-    out.push_back(std::vector<double>{0.3,0.0,0.6,0.8667});
-    out.push_back(std::vector<double>{0.35,0.0,0.6667,0.6667});
-    out.push_back(std::vector<double>{0.4,0.0,0.6667,0.5333});
-    out.push_back(std::vector<double>{0.45,0.0,0.6,0.0});
-    out.push_back(std::vector<double>{0.5,0.0,0.7333,0.0});
-    out.push_back(std::vector<double>{0.55,0.0,0.8667,0.0});
-    out.push_back(std::vector<double>{0.6,0.0,1.0,0.0});
-    out.push_back(std::vector<double>{0.65,0.7333,1.0,0.0});
-    out.push_back(std::vector<double>{0.7,0.9333,0.9333,0.0});
-    out.push_back(std::vector<double>{0.75,1.0,0.8,0.0});
-    out.push_back(std::vector<double>{0.8,1.0,0.6,0.0});
-    out.push_back(std::vector<double>{0.85,1.0,0.0,0.0});
-    out.push_back(std::vector<double>{0.9,0.8667,0.0,0.0});
-    out.push_back(std::vector<double>{0.95,0.8,0.0,0.0});
-    out.push_back(std::vector<double>{1.0,0.8,0.8,0.8});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_jet(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0,0,0.5});
-    out.push_back(std::vector<double>{0.11,0.0,0.0,1});
-    out.push_back(std::vector<double>{0.125,0.0,0,1.0});
-    out.push_back(std::vector<double>{0.34,0.0,0.86,1});
-    out.push_back(std::vector<double>{0.35,0,0.9,0.967741935484});
-    out.push_back(std::vector<double>{0.375,0.0806451612903,1,0.887096774194});
-    out.push_back(std::vector<double>{0.64,0.935483870968,1,0.0322580645161});
-    out.push_back(std::vector<double>{0.65,0.967741935484,0.962962962963,0});
-    out.push_back(std::vector<double>{0.66,1,0.925925925926,0.0});
-    out.push_back(std::vector<double>{0.89,1,0.0740740740741,0.0});
-    out.push_back(std::vector<double>{0.91,0.909090909091,0,0.0});
-    out.push_back(std::vector<double>{1,0.5,0,0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_rainbow(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,0.5,0.0,1.0});
-    out.push_back(std::vector<double>{0.01,0.48,0.0314107590781,0.999876632482});
-    out.push_back(std::vector<double>{0.02,0.46,0.0627905195293,0.999506560366});
-    out.push_back(std::vector<double>{0.03,0.44,0.0941083133185,0.998889874962});
-    out.push_back(std::vector<double>{0.04,0.42,0.125333233564,0.998026728428});
-    out.push_back(std::vector<double>{0.05,0.4,0.15643446504,0.996917333733});
-    out.push_back(std::vector<double>{0.06,0.38,0.187381314586,0.995561964603});
-    out.push_back(std::vector<double>{0.07,0.36,0.218143241397,0.993960955455});
-    out.push_back(std::vector<double>{0.08,0.34,0.248689887165,0.992114701314});
-    out.push_back(std::vector<double>{0.09,0.32,0.278991106039,0.990023657717});
-    out.push_back(std::vector<double>{0.1,0.3,0.309016994375,0.987688340595});
-    out.push_back(std::vector<double>{0.11,0.28,0.338737920245,0.985109326155});
-    out.push_back(std::vector<double>{0.12,0.26,0.368124552685,0.982287250729});
-    out.push_back(std::vector<double>{0.13,0.24,0.397147890635,0.979222810622});
-    out.push_back(std::vector<double>{0.14,0.22,0.425779291565,0.975916761939});
-    out.push_back(std::vector<double>{0.15,0.2,0.45399049974,0.972369920398});
-    out.push_back(std::vector<double>{0.16,0.18,0.481753674102,0.968583161129});
-    out.push_back(std::vector<double>{0.17,0.16,0.50904141575,0.964557418458});
-    out.push_back(std::vector<double>{0.18,0.14,0.535826794979,0.960293685677});
-    out.push_back(std::vector<double>{0.19,0.12,0.562083377852,0.955793014798});
-    out.push_back(std::vector<double>{0.2,0.1,0.587785252292,0.951056516295});
-    out.push_back(std::vector<double>{0.21,0.08,0.612907053653,0.946085358828});
-    out.push_back(std::vector<double>{0.22,0.06,0.637423989749,0.940880768954});
-    out.push_back(std::vector<double>{0.23,0.04,0.661311865324,0.93544403083});
-    out.push_back(std::vector<double>{0.24,0.02,0.684547105929,0.929776485888});
-    out.push_back(std::vector<double>{0.25,0.0,0.707106781187,0.923879532511});
-    out.push_back(std::vector<double>{0.26,0.02,0.728968627421,0.917754625684});
-    out.push_back(std::vector<double>{0.27,0.04,0.75011106963,0.911403276635});
-    out.push_back(std::vector<double>{0.28,0.06,0.770513242776,0.904827052466});
-    out.push_back(std::vector<double>{0.29,0.08,0.790155012376,0.898027575761});
-    out.push_back(std::vector<double>{0.3,0.1,0.809016994375,0.891006524188});
-    out.push_back(std::vector<double>{0.31,0.12,0.827080574275,0.883765630089});
-    out.push_back(std::vector<double>{0.32,0.14,0.844327925502,0.876306680044});
-    out.push_back(std::vector<double>{0.33,0.16,0.860742027004,0.868631514438});
-    out.push_back(std::vector<double>{0.34,0.18,0.876306680044,0.860742027004});
-    out.push_back(std::vector<double>{0.35,0.2,0.891006524188,0.852640164354});
-    out.push_back(std::vector<double>{0.36,0.22,0.904827052466,0.844327925502});
-    out.push_back(std::vector<double>{0.37,0.24,0.917754625684,0.835807361368});
-    out.push_back(std::vector<double>{0.38,0.26,0.929776485888,0.827080574275});
-    out.push_back(std::vector<double>{0.39,0.28,0.940880768954,0.818149717425});
-    out.push_back(std::vector<double>{0.4,0.3,0.951056516295,0.809016994375});
-    out.push_back(std::vector<double>{0.41,0.32,0.960293685677,0.799684658487});
-    out.push_back(std::vector<double>{0.42,0.34,0.968583161129,0.790155012376});
-    out.push_back(std::vector<double>{0.43,0.36,0.975916761939,0.780430407338});
-    out.push_back(std::vector<double>{0.44,0.38,0.982287250729,0.770513242776});
-    out.push_back(std::vector<double>{0.45,0.4,0.987688340595,0.7604059656});
-    out.push_back(std::vector<double>{0.46,0.42,0.992114701314,0.75011106963});
-    out.push_back(std::vector<double>{0.47,0.44,0.995561964603,0.739631094979});
-    out.push_back(std::vector<double>{0.48,0.46,0.998026728428,0.728968627421});
-    out.push_back(std::vector<double>{0.49,0.48,0.999506560366,0.718126297763});
-    out.push_back(std::vector<double>{0.5,0.5,1.0,0.707106781187});
-    out.push_back(std::vector<double>{0.51,0.52,0.999506560366,0.695912796592});
-    out.push_back(std::vector<double>{0.52,0.54,0.998026728428,0.684547105929});
-    out.push_back(std::vector<double>{0.53,0.56,0.995561964603,0.67301251351});
-    out.push_back(std::vector<double>{0.54,0.58,0.992114701314,0.661311865324});
-    out.push_back(std::vector<double>{0.55,0.6,0.987688340595,0.64944804833});
-    out.push_back(std::vector<double>{0.56,0.62,0.982287250729,0.637423989749});
-    out.push_back(std::vector<double>{0.57,0.64,0.975916761939,0.625242656336});
-    out.push_back(std::vector<double>{0.58,0.66,0.968583161129,0.612907053653});
-    out.push_back(std::vector<double>{0.59,0.68,0.960293685677,0.600420225326});
-    out.push_back(std::vector<double>{0.6,0.7,0.951056516295,0.587785252292});
-    out.push_back(std::vector<double>{0.61,0.72,0.940880768954,0.575005252043});
-    out.push_back(std::vector<double>{0.62,0.74,0.929776485888,0.562083377852});
-    out.push_back(std::vector<double>{0.63,0.76,0.917754625684,0.549022817998});
-    out.push_back(std::vector<double>{0.64,0.78,0.904827052466,0.535826794979});
-    out.push_back(std::vector<double>{0.65,0.8,0.891006524188,0.522498564716});
-    out.push_back(std::vector<double>{0.66,0.82,0.876306680044,0.50904141575});
-    out.push_back(std::vector<double>{0.67,0.84,0.860742027004,0.495458668432});
-    out.push_back(std::vector<double>{0.68,0.86,0.844327925502,0.481753674102});
-    out.push_back(std::vector<double>{0.69,0.88,0.827080574275,0.467929814261});
-    out.push_back(std::vector<double>{0.7,0.9,0.809016994375,0.45399049974});
-    out.push_back(std::vector<double>{0.71,0.92,0.790155012376,0.439939169856});
-    out.push_back(std::vector<double>{0.72,0.94,0.770513242776,0.425779291565});
-    out.push_back(std::vector<double>{0.73,0.96,0.75011106963,0.411514358605});
-    out.push_back(std::vector<double>{0.74,0.98,0.728968627421,0.397147890635});
-    out.push_back(std::vector<double>{0.75,1.0,0.707106781187,0.382683432365});
-    out.push_back(std::vector<double>{0.76,1.02,0.684547105929,0.368124552685});
-    out.push_back(std::vector<double>{0.77,1.04,0.661311865324,0.353474843779});
-    out.push_back(std::vector<double>{0.78,1.06,0.637423989749,0.338737920245});
-    out.push_back(std::vector<double>{0.79,1.08,0.612907053653,0.323917418198});
-    out.push_back(std::vector<double>{0.8,1.1,0.587785252292,0.309016994375});
-    out.push_back(std::vector<double>{0.81,1.12,0.562083377852,0.294040325232});
-    out.push_back(std::vector<double>{0.82,1.14,0.535826794979,0.278991106039});
-    out.push_back(std::vector<double>{0.83,1.16,0.50904141575,0.263873049965});
-    out.push_back(std::vector<double>{0.84,1.18,0.481753674102,0.248689887165});
-    out.push_back(std::vector<double>{0.85,1.2,0.45399049974,0.233445363856});
-    out.push_back(std::vector<double>{0.86,1.22,0.425779291565,0.218143241397});
-    out.push_back(std::vector<double>{0.87,1.24,0.397147890635,0.202787295357});
-    out.push_back(std::vector<double>{0.88,1.26,0.368124552685,0.187381314586});
-    out.push_back(std::vector<double>{0.89,1.28,0.338737920245,0.171929100279});
-    out.push_back(std::vector<double>{0.9,1.3,0.309016994375,0.15643446504});
-    out.push_back(std::vector<double>{0.91,1.32,0.278991106039,0.140901231938});
-    out.push_back(std::vector<double>{0.92,1.34,0.248689887165,0.125333233564});
-    out.push_back(std::vector<double>{0.93,1.36,0.218143241397,0.109734311091});
-    out.push_back(std::vector<double>{0.94,1.38,0.187381314586,0.0941083133185});
-    out.push_back(std::vector<double>{0.95,1.4,0.15643446504,0.0784590957278});
-    out.push_back(std::vector<double>{0.96,1.42,0.125333233564,0.0627905195293});
-    out.push_back(std::vector<double>{0.97,1.44,0.0941083133185,0.0471064507096});
-    out.push_back(std::vector<double>{0.98,1.46,0.0627905195293,0.0314107590781});
-    out.push_back(std::vector<double>{0.99,1.48,0.0314107590781,0.0157073173118});
-    out.push_back(std::vector<double>{1.0,1.5,1.22464679915e-16,6.12323399574e-17});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_gist_rainbow(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,0.0,0.16,1.0});
-    out.push_back(std::vector<double>{0.03,1.0,0.0,0.0,1.0});
-    out.push_back(std::vector<double>{0.215,1.0,1.0,0.0,1.0});
-    out.push_back(std::vector<double>{0.4,0.0,1.0,0.0,1.0});
-    out.push_back(std::vector<double>{0.586,0.0,1.0,1.0,1.0});
-    out.push_back(std::vector<double>{0.77,0.0,0.0,1.0,1.0});
-    out.push_back(std::vector<double>{0.954,1.0,0.0,1.0,1.0});
-    out.push_back(std::vector<double>{1.0,1.0,0.0,0.75,1.0});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_hsv(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.0,0.0,0.0});
-    out.push_back(std::vector<double>{0.15873,1.0,0.9375,0.0});
-    out.push_back(std::vector<double>{0.174603,0.96875,1.0,0.0});
-    out.push_back(std::vector<double>{0.333333,0.03125,1.0,0.0});
-    out.push_back(std::vector<double>{0.349206,0.0,1.0,0.0625});
-    out.push_back(std::vector<double>{0.507937,0.0,1.0,1.0});
-    out.push_back(std::vector<double>{0.666667,0.0,0.0625,1.0});
-    out.push_back(std::vector<double>{0.68254,0.03125,0.0,1.0});
-    out.push_back(std::vector<double>{0.84127,0.96875,0.0,1.0});
-    out.push_back(std::vector<double>{0.857143,1.0,0.0,0.9375});
-    out.push_back(std::vector<double>{1.0,1.0,0.0,0.09375});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_flag(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.03033008589,0.0,-0.0303300858899});
-    out.push_back(std::vector<double>{0.01,1.23441710797,0.835807361368,0.652090471517});
-    out.push_back(std::vector<double>{0.02,0.776093414514,0.917754625684,1.19733236442});
-    out.push_back(std::vector<double>{0.03,0.0687460609675,0.171929100279,1.11361228807});
-    out.push_back(std::vector<double>{0.04,-0.249629920274,-0.728968627421,0.476441930691});
-    out.push_back(std::vector<double>{0.05,0.108126076463,-0.972369920398,-0.139480123266});
-    out.push_back(std::vector<double>{0.06,0.819334468674,-0.338737920245,-0.17862028935});
-    out.push_back(std::vector<double>{0.07,1.24251774329,0.600420225326,0.394324076047});
-    out.push_back(std::vector<double>{0.08,0.995983898993,0.998026728428,1.06258330222});
-    out.push_back(std::vector<double>{0.09,0.302095212526,0.495458668432,1.22341806384});
-    out.push_back(std::vector<double>{0.1,-0.213292387221,-0.45399049974,0.731762745781});
-    out.push_back(std::vector<double>{0.11,-0.0853228055037,-0.993960955455,0.0310680077482});
-    out.push_back(std::vector<double>{0.12,0.570581234989,-0.637423989749,-0.246671473452});
-    out.push_back(std::vector<double>{0.13,1.16282422257,0.294040325232,0.149052639305});
-    out.push_back(std::vector<double>{0.14,1.15723001003,0.960293685677,0.861315255576});
-    out.push_back(std::vector<double>{0.15,0.558844321796,0.7604059656,1.2476880003});
-    out.push_back(std::vector<double>{0.16,-0.0926162592818,-0.125333233564,0.95968029024});
-    out.push_back(std::vector<double>{0.17,-0.209564019121,-0.898027575761,0.257061936351});
-    out.push_back(std::vector<double>{0.18,0.313482584626,-0.860742027004,-0.226437370846});
-    out.push_back(std::vector<double>{0.19,1.00475938513,-0.0471064507096,-0.054723321234});
-    out.push_back(std::vector<double>{0.2,1.24076625545,0.809016994375,0.61732584878});
-    out.push_back(std::vector<double>{0.21,0.808635768954,0.93544403083,1.18355245748});
-    out.push_back(std::vector<double>{0.22,0.0981299037658,0.218143241397,1.13324594413});
-    out.push_back(std::vector<double>{0.23,-0.249907474361,-0.695912796592,0.511780487984});
-    out.push_back(std::vector<double>{0.24,0.0784374666109,-0.982287250729,-0.120310430706});
-    out.push_back(std::vector<double>{0.25,0.787012574274,-0.382683432365,-0.192909649383});
-    out.push_back(std::vector<double>{0.26,1.23671543805,0.562083377852,0.359464014061});
-    out.push_back(std::vector<double>{0.27,1.02193459744,0.999876632482,1.03859472332});
-    out.push_back(std::vector<double>{0.28,0.336392568953,0.535826794979,1.23193757145});
-    out.push_back(std::vector<double>{0.29,-0.201583023122,-0.411514358605,0.765106132834});
-    out.push_back(std::vector<double>{0.3,-0.106762745781,-0.987688340595,0.0591610607806});
-    out.push_back(std::vector<double>{0.31,0.535329838032,-0.67301251351,-0.249167406221});
-    out.push_back(std::vector<double>{0.32,1.14555652025,0.248689887165,0.118218938187});
-    out.push_back(std::vector<double>{0.33,1.17352068182,0.946085358828,0.829954377392});
-    out.push_back(std::vector<double>{0.34,0.593999925173,0.790155012376,1.24408602599});
-    out.push_back(std::vector<double>{0.35,-0.0703044742,-0.0784590957278,0.987086036248});
-    out.push_back(std::vector<double>{0.36,-0.220220264258,-0.876306680044,0.290756670471});
-    out.push_back(std::vector<double>{0.37,0.279469756076,-0.883765630089,-0.216844761099});
-    out.push_back(std::vector<double>{0.38,0.978067992312,-0.0941083133185,-0.0778849320818});
-    out.push_back(std::vector<double>{0.39,1.24547071659,0.780430407338,0.582300733318});
-    out.push_back(std::vector<double>{0.4,0.840492874805,0.951056516295,1.16825489314});
-    out.push_back(std::vector<double>{0.41,0.128405998676,0.263873049965,1.15147363583});
-    out.push_back(std::vector<double>{0.42,-0.248520046321,-0.661311865324,0.547092889647});
-    out.push_back(std::vector<double>{0.43,0.0496848310056,-0.990023657717,-0.0997634938653});
-    out.push_back(std::vector<double>{0.44,0.754053440184,-0.425779291565,-0.205660576716});
-    out.push_back(std::vector<double>{0.45,1.2292774403,0.522498564716,0.324915977108});
-    out.push_back(std::vector<double>{0.46,1.04672647057,0.999506560366,1.01341032945});
-    out.push_back(std::vector<double>{0.47,0.37105317479,0.575005252043,1.23883199462});
-    out.push_back(std::vector<double>{0.48,-0.188315969263,-0.368124552685,0.797860917976});
-    out.push_back(std::vector<double>{0.49,-0.126855521026,-0.979222810622,0.0882328865014});
-    out.push_back(std::vector<double>{0.5,0.5,-0.707106781187,-0.25});
-    out.push_back(std::vector<double>{0.51,1.12685552103,0.202787295357,0.0882328865014});
-    out.push_back(std::vector<double>{0.52,1.18831596926,0.929776485888,0.797860917976});
-    out.push_back(std::vector<double>{0.53,0.62894682521,0.818149717425,1.23883199462});
-    out.push_back(std::vector<double>{0.54,-0.0467264705661,-0.0314107590781,1.01341032945});
-    out.push_back(std::vector<double>{0.55,-0.229277440298,-0.852640164354,0.324915977108});
-    out.push_back(std::vector<double>{0.56,0.245946559816,-0.904827052466,-0.205660576716});
-    out.push_back(std::vector<double>{0.57,0.950315168994,-0.140901231938,-0.0997634938653});
-    out.push_back(std::vector<double>{0.58,1.24852004632,0.75011106963,0.547092889647});
-    out.push_back(std::vector<double>{0.59,0.871594001324,0.964557418458,1.15147363583});
-    out.push_back(std::vector<double>{0.6,0.159507125195,0.309016994375,1.16825489314});
-    out.push_back(std::vector<double>{0.61,-0.245470716591,-0.625242656336,0.582300733318});
-    out.push_back(std::vector<double>{0.62,0.0219320076885,-0.995561964603,-0.0778849320818});
-    out.push_back(std::vector<double>{0.63,0.720530243924,-0.467929814261,-0.216844761099});
-    out.push_back(std::vector<double>{0.64,1.22022026426,0.481753674102,0.290756670471});
-    out.push_back(std::vector<double>{0.65,1.0703044742,0.996917333733,0.987086036248});
-    out.push_back(std::vector<double>{0.66,0.406000074827,0.612907053653,1.24408602599});
-    out.push_back(std::vector<double>{0.67,-0.17352068182,-0.323917418198,0.829954377392});
-    out.push_back(std::vector<double>{0.68,-0.145556520253,-0.968583161129,0.118218938187});
-    out.push_back(std::vector<double>{0.69,0.464670161968,-0.739631094979,-0.249167406221});
-    out.push_back(std::vector<double>{0.7,1.10676274578,0.15643446504,0.0591610607806});
-    out.push_back(std::vector<double>{0.71,1.20158302312,0.911403276635,0.765106132834});
-    out.push_back(std::vector<double>{0.72,0.663607431047,0.844327925502,1.23193757145});
-    out.push_back(std::vector<double>{0.73,-0.0219345974442,0.0157073173118,1.03859472332});
-    out.push_back(std::vector<double>{0.74,-0.236715438047,-0.827080574275,0.359464014061});
-    out.push_back(std::vector<double>{0.75,0.212987425726,-0.923879532511,-0.192909649383});
-    out.push_back(std::vector<double>{0.76,0.921562533389,-0.187381314586,-0.120310430706});
-    out.push_back(std::vector<double>{0.77,1.24990747436,0.718126297763,0.511780487984});
-    out.push_back(std::vector<double>{0.78,0.901870096234,0.975916761939,1.13324594413});
-    out.push_back(std::vector<double>{0.79,0.191364231046,0.353474843779,1.18355245748});
-    out.push_back(std::vector<double>{0.8,-0.240766255446,-0.587785252292,0.61732584878});
-    out.push_back(std::vector<double>{0.81,-0.00475938513233,-0.998889874962,-0.054723321234});
-    out.push_back(std::vector<double>{0.82,0.686517415374,-0.50904141575,-0.226437370846});
-    out.push_back(std::vector<double>{0.83,1.20956401912,0.439939169856,0.257061936351});
-    out.push_back(std::vector<double>{0.84,1.09261625928,0.992114701314,0.95968029024});
-    out.push_back(std::vector<double>{0.85,0.441155678204,0.64944804833,1.2476880003});
-    out.push_back(std::vector<double>{0.86,-0.157230010033,-0.278991106039,0.861315255576});
-    out.push_back(std::vector<double>{0.87,-0.162824222567,-0.955793014798,0.149052639305});
-    out.push_back(std::vector<double>{0.88,0.429418765011,-0.770513242776,-0.246671473452});
-    out.push_back(std::vector<double>{0.89,1.0853228055,0.109734311091,0.0310680077482});
-    out.push_back(std::vector<double>{0.9,1.21329238722,0.891006524188,0.731762745781});
-    out.push_back(std::vector<double>{0.91,0.697904787474,0.868631514438,1.22341806384});
-    out.push_back(std::vector<double>{0.92,0.00401610100726,0.0627905195293,1.06258330222});
-    out.push_back(std::vector<double>{0.93,-0.242517743287,-0.799684658487,0.394324076047});
-    out.push_back(std::vector<double>{0.94,0.180665531326,-0.940880768954,-0.17862028935});
-    out.push_back(std::vector<double>{0.95,0.891873923537,-0.233445363856,-0.139480123266});
-    out.push_back(std::vector<double>{0.96,1.24962992027,0.684547105929,0.476441930691});
-    out.push_back(std::vector<double>{0.97,0.931253939032,0.985109326155,1.11361228807});
-    out.push_back(std::vector<double>{0.98,0.223906585486,0.397147890635,1.19733236442});
-    out.push_back(std::vector<double>{0.99,-0.234417107966,-0.549022817998,0.652090471517});
-    out.push_back(std::vector<double>{1.0,-0.0303300858899,-1.0,-0.0303300858899});
-    return out;
-}
-
-static std::vector<std::vector<double>> make_color_map_prism(void)
-{
-    std::vector<std::vector<double>> out;
-    out.push_back(std::vector<double>{0.0,1.20033008589,-0.20033008589,-0.0});
-    out.push_back(std::vector<double>{0.01,1.41378704508,0.23366292735,-0.67146385678});
-    out.push_back(std::vector<double>{0.02,1.31794256289,0.707717401227,-1.06370163654});
-    out.push_back(std::vector<double>{0.03,0.952653257809,1.02469931326,-1.0136025249});
-    out.push_back(std::vector<double>{0.04,0.46982350801,1.05279275872,-0.542000075703});
-    out.push_back(std::vector<double>{0.05,0.0702365061347,0.780315168994,0.154991355131});
-    out.push_back(std::vector<double>{0.06,-0.0799407831529,0.320575470087,0.787530131324});
-    out.push_back(std::vector<double>{0.07,0.0817423810499,-0.135245068482,1.09257704495});
-    out.push_back(std::vector<double>{0.08,0.488050576903,-0.397594947368,0.943279321813});
-    out.push_back(std::vector<double>{0.09,0.970021852928,-0.35737681643,0.401721938081});
-    out.push_back(std::vector<double>{0.1,1.32723001003,-0.0313152555763,-0.306890216643});
-    out.push_back(std::vector<double>{0.11,1.41113118933,0.444998087803,-0.887883065571});
-    out.push_back(std::vector<double>{0.12,1.18683535629,0.87348984764,-1.09965260821});
-    out.push_back(std::vector<double>{0.13,0.747615208938,1.07597310899,-0.854135163956});
-    out.push_back(std::vector<double>{0.14,0.276119027529,0.968245861346,-0.253428369344});
-    out.push_back(std::vector<double>{0.15,-0.0315830231224,0.595106132834,0.452665794466});
-    out.push_back(std::vector<double>{0.16,-0.0475340386691,0.111722874879,0.970520349078});
-    out.push_back(std::vector<double>{0.17,0.234899159206,-0.280890545303,1.08478714985});
-    out.push_back(std::vector<double>{0.18,0.698267637002,-0.41946710448,0.747948716595});
-    out.push_back(std::vector<double>{0.19,1.14988110921,-0.246380187914,0.100078224294});
-    out.push_back(std::vector<double>{0.2,1.40193757145,0.166392568953,-0.589409474477});
-    out.push_back(std::vector<double>{0.21,1.34962015767,0.647200947814,-1.03379314619});
-    out.push_back(std::vector<double>{0.22,1.01468489547,0.99610233661,-1.04827727567});
-    out.push_back(std::vector<double>{0.23,0.536413532302,1.06800721924,-0.626838684419});
-    out.push_back(std::vector<double>{0.24,0.113693670446,0.833014182403,0.0552687499977});
-    out.push_back(std::vector<double>{0.25,-0.0776880002998,0.388844321796,0.714392853163});
-    out.push_back(std::vector<double>{0.26,0.0418539699684,-0.0797957600507,1.07643899476});
-    out.push_back(std::vector<double>{0.27,0.422608438856,-0.378023598106,0.990851455666});
-    out.push_back(std::vector<double>{0.28,0.906239889741,-0.381822108743,0.493221537699});
-    out.push_back(std::vector<double>{0.29,1.29163174518,-0.0896116935766,-0.209512964354});
-    out.push_back(std::vector<double>{0.3,1.41852004632,0.377092889647,-0.825122176594});
-    out.push_back(std::vector<double>{0.31,1.2341387028,0.82421404675,-1.0976069956});
-    out.push_back(std::vector<double>{0.32,0.815162101038,1.06581788808,-0.913655489115});
-    out.push_back(std::vector<double>{0.33,0.335820308217,1.00143423624,-0.349763297216});
-    out.push_back(std::vector<double>{0.34,-0.00455393867478,0.657836824988,0.35957683262});
-    out.push_back(std::vector<double>{0.35,-0.0644171079663,0.177909528483,0.919388097505});
-    out.push_back(std::vector<double>{0.36,0.181124706024,-0.238771436467,1.09687479029});
-    out.push_back(std::vector<double>{0.37,0.629963732681,-0.41893063584,0.818229642788});
-    out.push_back(std::vector<double>{0.38,1.09545171185,-0.287649448221,0.19932633999});
-    out.push_back(std::vector<double>{0.39,1.38401697021,0.100479268359,-0.502466177768});
-    out.push_back(std::vector<double>{0.4,1.37566057672,0.584053440184,-0.995309757713});
-    out.push_back(std::vector<double>{0.41,1.07385751062,0.96198030912,-1.07425698827});
-    out.push_back(std::vector<double>{0.42,0.604111602587,1.07710020686,-0.706477918494});
-    out.push_back(std::vector<double>{0.43,0.161765171897,0.881540895586,-0.0449122871764});
-    out.push_back(std::vector<double>{0.44,-0.0692334463617,0.456625083535,0.635329973764});
-    out.push_back(std::vector<double>{0.45,0.00717577743348,-0.0209473606954,1.05137231628});
-    out.push_back(std::vector<double>{0.46,0.359218314255,-0.352579478014,1.03020487542});
-    out.push_back(std::vector<double>{0.47,0.840498411735,-0.400363123108,0.58063006314});
-    out.push_back(std::vector<double>{0.48,1.25087729515,-0.144427621435,-0.110397886336});
-    out.push_back(std::vector<double>{0.49,1.41970023074,0.308797074922,-0.755517231199});
-    out.push_back(std::vector<double>{0.5,1.27676274578,0.770838939219,-1.08645717465});
-    out.push_back(std::vector<double>{0.51,0.881504931925,1.04955935389,-0.965597409657});
-    out.push_back(std::vector<double>{0.52,0.398293474972,1.02905333434,-0.443197079285});
-    out.push_back(std::vector<double>{0.53,0.0280702992467,0.717848242604,0.263505325646});
-    out.push_back(std::vector<double>{0.54,-0.07520848289,0.245357711345,0.860629891634});
-    out.push_back(std::vector<double>{0.55,0.131405276678,-0.191934597444,1.09986429573});
-    out.push_back(std::vector<double>{0.56,0.561991913086,-0.412182088952,0.881723683358});
-    out.push_back(std::vector<double>{0.57,1.03749336373,-0.323795554904,0.296921123907});
-    out.push_back(std::vector<double>{0.58,1.36017388552,0.0364697498721,-0.411355126312});
-    out.push_back(std::vector<double>{0.59,1.39584782525,0.518798661501,-0.948570675313});
-    out.push_back(std::vector<double>{0.6,1.12968029024,0.922616259282,-1.09132617145});
-    out.push_back(std::vector<double>{0.61,0.672356190614,1.0799962989,-0.78025720253});
-    out.push_back(std::vector<double>{0.62,0.214052276729,0.925492798986,-0.144720795002});
-    out.push_back(std::vector<double>{0.63,-0.0546472484613,0.523355541161,0.550997287897});
-    out.push_back(std::vector<double>{0.64,-0.0220045549026,0.040812005792,1.01758492752});
-    out.push_back(std::vector<double>{0.65,0.298405998676,-0.321473635829,1.0610131603});
-    out.push_back(std::vector<double>{0.66,0.773342718013,-0.412846069272,0.663222495858});
-    out.push_back(std::vector<double>{0.67,1.20530470171,-0.195308363088,-0.0103671022765});
-    out.push_back(std::vector<double>{0.68,1.41466195343,0.240677129929,-0.679645574399});
-    out.push_back(std::vector<double>{0.69,1.31435393596,0.713807250086,-1.06629562864});
-    out.push_back(std::vector<double>{0.7,0.946093414514,1.02733236442,-1.00953008825});
-    out.push_back(std::vector<double>{0.71,0.463020338013,1.05087406634,-0.532954719929});
-    out.push_back(std::vector<double>{0.72,0.0660190857167,0.774642615121,0.165248148033});
-    out.push_back(std::vector<double>{0.73,-0.0798186533277,0.313507967931,0.794733107139});
-    out.push_back(std::vector<double>{0.74,0.0861532738247,-0.140768520968,1.09373086944});
-    out.push_back(std::vector<double>{0.75,0.494915977108,-0.399277440298,0.93790418079});
-    out.push_back(std::vector<double>{0.76,0.976486806052,-0.354518690553,0.392053066585});
-    out.push_back(std::vector<double>{0.77,1.33060608619,-0.0251050533092,-0.316832049616});
-    out.push_back(std::vector<double>{0.78,1.41001445816,0.451977873896,-0.893963580876});
-    out.push_back(std::vector<double>{0.79,1.18169020683,0.878336696047,-1.09934324319});
-    out.push_back(std::vector<double>{0.8,0.740581234989,1.07667147345,-0.847564567053});
-    out.push_back(std::vector<double>{0.81,0.270121284447,0.964505329251,-0.243328902069});
-    out.push_back(std::vector<double>{0.82,-0.0340503932404,0.588482192381,0.462094301253});
-    out.push_back(std::vector<double>{0.83,-0.0454449879255,0.104970070319,0.975357081271});
-    out.push_back(std::vector<double>{0.84,0.240675905804,-0.284964081994,1.08302076798});
-    out.push_back(std::vector<double>{0.85,0.705329838032,-0.419167406221,0.740313764861});
-    out.push_back(std::vector<double>{0.86,1.15529197118,-0.241831883259,0.089749672725});
-    out.push_back(std::vector<double>{0.87,1.40344700487,0.173298082198,-0.598136531057});
-    out.push_back(std::vector<double>{0.88,1.34660046953,0.653592034261,-1.03728958948});
-    out.push_back(std::vector<double>{0.89,1.00839181333,0.999321283593,-1.04508912101});
-    out.push_back(std::vector<double>{0.9,0.529464014061,1.06671543805,-0.618291715637});
-    out.push_back(std::vector<double>{0.91,0.108977650822,0.827748856074,0.0656203035292});
-    out.push_back(std::vector<double>{0.92,-0.0782093797709,0.381795019286,0.722244331328});
-    out.push_back(std::vector<double>{0.93,0.0457440446195,-0.0856976090525,1.07852538571});
-    out.push_back(std::vector<double>{0.94,0.429292292645,-0.380323728746,0.986305140464});
-    out.push_back(std::vector<double>{0.95,0.912938063649,-0.379564019121,0.483933086842});
-    out.push_back(std::vector<double>{0.96,1.29555882521,-0.0837344029713,-0.219680978566});
-    out.push_back(std::vector<double>{0.97,1.41804296892,0.384145328912,-0.831941418277});
-    out.push_back(std::vector<double>{0.98,1.22945585907,0.829508900576,-1.09824170512});
-    out.push_back(std::vector<double>{0.99,0.808220836422,1.0671533086,-0.907841724562});
-    out.push_back(std::vector<double>{1.0,0.329507125195,0.998254893141,-0.339918693812});
-    return out;
-}
-
-
-static const char *COLOR_MAP_TITLES[] = {
-    "Sequential/Blues",
-    "Sequential/Bugn",
-    "Sequential/Bupu",
-    "Sequential/Gnbu",
-    "Sequential/Greens",
-    "Sequential/Greys",
-    "Sequential/Oranges",
-    "Sequential/Orrd",
-    "Sequential/Pubu",
-    "Sequential/Pubugn",
-    "Sequential/Purd",
-    "Sequential/Purples",
-    "Sequential/Rdpu",
-    "Sequential/Reds",
-    "Sequential/Ylgn",
-    "Sequential/Ylgnbu",
-    "Sequential/Ylorbr",
-    "Sequential/Ylorrd",
-    "Sequential (2)/Afmhot",
-    "Sequential (2)/Autumn",
-    "Sequential (2)/Bone",
-    "Sequential (2)/Cool",
-    "Sequential (2)/Copper",
-    "Sequential (2)/Gist Heat",
-    "Sequential (2)/Gray",
-    "Sequential (2)/Hot",
-    "Sequential (2)/Pink",
-    "Sequential (2)/Spring",
-    "Sequential (2)/Summer",
-    "Sequential (2)/Winter",
-    "Diverging/Brbg",
-    "Diverging/Bwr",
-    "Diverging/Coolwarm",
-    "Diverging/Piyg",
-    "Diverging/Prgn",
-    "Diverging/Puor",
-    "Diverging/Rdbu",
-    "Diverging/Rdgy",
-    "Diverging/Rdylbu",
-    "Diverging/Rdylgn",
-    "Diverging/Spectral",
-    "Diverging/Seismic",
-    "Qualitative/Accent",
-    "Qualitative/Dark2",
-    "Qualitative/Paired",
-    "Qualitative/Pastel1",
-    "Qualitative/Pastel2",
-    "Qualitative/Set1",
-    "Qualitative/Set2",
-    "Qualitative/Set3",
-    "Miscellaneous/Gist Earth",
-    "Miscellaneous/Terrain",
-    "Miscellaneous/Ocean",
-    "Miscellaneous/Gist Stern",
-    "Miscellaneous/Brg",
-    "Miscellaneous/Cmrmap",
-    "Miscellaneous/Cubehelix",
-    "Miscellaneous/Gnuplot",
-    "Miscellaneous/Gnuplot2",
-    "Miscellaneous/Gist Ncar",
-    "Miscellaneous/Nipy Spectral",
-    "Miscellaneous/Jet",
-    "Miscellaneous/Rainbow",
-    "Miscellaneous/Gist Rainbow",
-    "Miscellaneous/Hsv",
-    "Miscellaneous/Flag",
-    "Miscellaneous/Prism"};
-
-static const char *COLOR_MAP_NAMES[] = {
-    "Blues",
-    "BuGn",
-    "BuPu",
-    "GnBu",
-    "Greens",
-    "Greys",
-    "Oranges",
-    "OrRd",
-    "PuBu",
-    "PuBuGn",
-    "PuRd",
-    "Purples",
-    "RdPu",
-    "Reds",
-    "YlGn",
-    "YlGnBu",
-    "YlOrBr",
-    "YlOrRd",
-    "afmhot",
-    "autumn",
-    "bone",
-    "cool",
-    "copper",
-    "gist_heat",
-    "gray",
-    "hot",
-    "pink",
-    "spring",
-    "summer",
-    "winter",
-    "BrBG",
-    "bwr",
-    "coolwarm",
-    "PiYG",
-    "PRGn",
-    "PuOr",
-    "RdBu",
-    "RdGy",
-    "RdYlBu",
-    "RdYlGn",
-    "Spectral",
-    "seismic",
-    "Accent",
-    "Dark2",
-    "Paired",
-    "Pastel1",
-    "Pastel2",
-    "Set1",
-    "Set2",
-    "Set3",
-    "gist_earth",
-    "terrain",
-    "ocean",
-    "gist_stern",
-    "brg",
-    "CMRmap",
-    "cubehelix",
-    "gnuplot",
-    "gnuplot2",
-    "gist_ncar",
-    "nipy_spectral",
-    "jet",
-    "rainbow",
-    "gist_rainbow",
-    "hsv",
-    "flag",
-    "prism"};
-
-#define NUM_COLOR_MAPS 67
-
-static std::vector<std::vector<double>> make_color_map(const std::string &name)
-{
-    if (name == "Blues") return make_color_map_Blues();
-    if (name == "BuGn") return make_color_map_BuGn();
-    if (name == "BuPu") return make_color_map_BuPu();
-    if (name == "GnBu") return make_color_map_GnBu();
-    if (name == "Greens") return make_color_map_Greens();
-    if (name == "Greys") return make_color_map_Greys();
-    if (name == "Oranges") return make_color_map_Oranges();
-    if (name == "OrRd") return make_color_map_OrRd();
-    if (name == "PuBu") return make_color_map_PuBu();
-    if (name == "PuBuGn") return make_color_map_PuBuGn();
-    if (name == "PuRd") return make_color_map_PuRd();
-    if (name == "Purples") return make_color_map_Purples();
-    if (name == "RdPu") return make_color_map_RdPu();
-    if (name == "Reds") return make_color_map_Reds();
-    if (name == "YlGn") return make_color_map_YlGn();
-    if (name == "YlGnBu") return make_color_map_YlGnBu();
-    if (name == "YlOrBr") return make_color_map_YlOrBr();
-    if (name == "YlOrRd") return make_color_map_YlOrRd();
-    if (name == "afmhot") return make_color_map_afmhot();
-    if (name == "autumn") return make_color_map_autumn();
-    if (name == "bone") return make_color_map_bone();
-    if (name == "cool") return make_color_map_cool();
-    if (name == "copper") return make_color_map_copper();
-    if (name == "gist_heat") return make_color_map_gist_heat();
-    if (name == "gray") return make_color_map_gray();
-    if (name == "hot") return make_color_map_hot();
-    if (name == "pink") return make_color_map_pink();
-    if (name == "spring") return make_color_map_spring();
-    if (name == "summer") return make_color_map_summer();
-    if (name == "winter") return make_color_map_winter();
-    if (name == "BrBG") return make_color_map_BrBG();
-    if (name == "bwr") return make_color_map_bwr();
-    if (name == "coolwarm") return make_color_map_coolwarm();
-    if (name == "PiYG") return make_color_map_PiYG();
-    if (name == "PRGn") return make_color_map_PRGn();
-    if (name == "PuOr") return make_color_map_PuOr();
-    if (name == "RdBu") return make_color_map_RdBu();
-    if (name == "RdGy") return make_color_map_RdGy();
-    if (name == "RdYlBu") return make_color_map_RdYlBu();
-    if (name == "RdYlGn") return make_color_map_RdYlGn();
-    if (name == "Spectral") return make_color_map_Spectral();
-    if (name == "seismic") return make_color_map_seismic();
-    if (name == "Accent") return make_color_map_Accent();
-    if (name == "Dark2") return make_color_map_Dark2();
-    if (name == "Paired") return make_color_map_Paired();
-    if (name == "Pastel1") return make_color_map_Pastel1();
-    if (name == "Pastel2") return make_color_map_Pastel2();
-    if (name == "Set1") return make_color_map_Set1();
-    if (name == "Set2") return make_color_map_Set2();
-    if (name == "Set3") return make_color_map_Set3();
-    if (name == "gist_earth") return make_color_map_gist_earth();
-    if (name == "terrain") return make_color_map_terrain();
-    if (name == "ocean") return make_color_map_ocean();
-    if (name == "gist_stern") return make_color_map_gist_stern();
-    if (name == "brg") return make_color_map_brg();
-    if (name == "CMRmap") return make_color_map_CMRmap();
-    if (name == "cubehelix") return make_color_map_cubehelix();
-    if (name == "gnuplot") return make_color_map_gnuplot();
-    if (name == "gnuplot2") return make_color_map_gnuplot2();
-    if (name == "gist_ncar") return make_color_map_gist_ncar();
-    if (name == "nipy_spectral") return make_color_map_nipy_spectral();
-    if (name == "jet") return make_color_map_jet();
-    if (name == "rainbow") return make_color_map_rainbow();
-    if (name == "gist_rainbow") return make_color_map_gist_rainbow();
-    if (name == "hsv") return make_color_map_hsv();
-    if (name == "flag") return make_color_map_flag();
-    if (name == "prism") return make_color_map_prism();
-    return std::vector<std::vector<double>>();
-}
+//! Get a color mapping given a name in [Z, R, G, B, A] format
+std::vector<std::vector<double>> make_color_map(const std::string &name);
